+++ include/mathalea_exercices_3e.js
@@ -1,1927 +1,21 @@
 
 ;
 
-/**
- * 3F12 Notion de fonction - Vocabulaire
- * Déterminer à partir de plusieurs modes de représentation l'image d'un nombre
- * @author Sébastien LOZANO
- */
+;
 
-function fonctions_calculs_d_images() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.sup = 1;
-	this.titre = "Fonctions : Calculs d'images";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = ``;
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne += `Calcule les images avec la méthode demandée.`;
+;
 
-	sortie_html ? this.spacing = 2 : this.spacing = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	this.nb_questions = 4;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 5;
+;
 
-	var num_ex = '3F12'; // pour rendre unique les id des SVG, en cas d'utilisation dans plusieurs exercices y faisant appel
+;
 
-	if (sortie_html) {
-		var pourcentage = '100%'; // pour l'affichage des svg. On a besoin d'une variable globale
-	} else { // sortie LaTeX
+;
 
-	};
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//			 this.bouton_aide = modal_pdf(numero_de_l_exercice,"pdf/FicheFonctions-3F1-act.pdf","Aide mémoire sur les fonctions (Sébastien Lozano)","Aide mémoire")		
-			//			 this.bouton_aide += modal_video('videoTest','videos/Fonctions.mp4','Petit conte mathématique','Intro Vidéo');
-		}
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
+;
 
-		let type_de_questions_disponibles = [];
-		if (this.sup == 1) {
-			type_de_questions_disponibles = [1]; // prog de calcul
-		} else if (this.sup == 2) {
-			type_de_questions_disponibles = [2]; // diagramme
-		} else if (this.sup == 3) {
-			type_de_questions_disponibles = [3]; // f(x) = ...
-		} else if (this.sup == 4) {
-			type_de_questions_disponibles = [4]; // f : x ---> ...
-		} else if (this.sup == 5) {
-			type_de_questions_disponibles = [1, 2, 3, 4]; // mélange
-		};
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
+;
 
-		for (let i = 0, a, b, c, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-			if (sortie_html) {
-				let id_unique = `${num_ex}_${i}_${Date.now()}`
-				var id_du_div = `div_svg${numero_de_l_exercice}${id_unique}`;
-				var id_du_div_corr = `div_svg_corr${numero_de_l_exercice}${id_unique}`;
-			}
-			// on part sur de l'affine avec coeff positifs, on verra ensuite
-			a = randint(2, 9);
-			b = randint(2, 9);
-			c = randint(2, 9);
-
-			switch (type_de_questions) {
-				case 1:
-					var j = 0; // pour la sous-numérotation
-					texte = `On donne le programme de calcul suivant qui correspond à une certaine fonction :`;
-					texte_corr = `Avec ce programme de calcul :`
-					if (sortie_html) {
-						texte += `
-							<br>
-							<div class="ui compact warning message">		
-							<p>							
-							- Choisir un nombre<br>
-							- Multiplier ce nombre par ${a}<br>
-							- Ajouter ${b} au résultat obtenu<br>
-							</p>
-							</div>
-							<br>`;
-						// sous-question a/
-						texte += num_alpha(j) + ` Appliquer ce programme de calcul au nombre ${c}<br>`;
-						texte_corr += `<br>` + num_alpha(j) + `
-							<br>
-							<div class="ui compact warning message">		
-							<p>							
-							- On choisit le nombre ${c}<br>
-							- On multiplie ce nombre par ${a} : ${a}$\\times$ ${c} = ${a * c}<br>
-							- On ajoute ${b} au résultat obtenu : ${a * c}+${b}=${a * c + b}<br>
-							</p>
-							</div>
-							<br>							
-							`;
-						j++;
-						// sous-question b/
-						texte += num_alpha(j) + ` Traduire ce calcul par une phrase contenant le mot image`;
-						texte_corr += num_alpha(j) + `L'image de ${c} par cette fonction vaut ${a * c + b}`;
-						texte_corr += `<br> On peut aussi dire que ${a * c + b} est l'image de ${c} par cette fonction`;
-					} else {
-						texte += tex_cadre_par_orange(itemize([`Choisir un nombre`, `Multiplier ce nombre par ${a}`, `Ajouter ${b} au résultat obtenu`]));
-						// sous-question a/
-						texte += tex_enumerate([`Appliquer ce programme de calcul au nombre ${c}`, `Traduire ce calcul par une phrase contenant le mot image`], this.spacing);
-						//texte_corr += 
-						texte_corr += tex_enumerate([tex_cadre_par_orange(itemize([`On choisit le nombre ${c}`, `On multiplie ce nombre par ${a} : $${a} \\times ${c} = ${a * c}$ `, `On ajoute ${b} au résultat obtenu : $${a * c}+${b}=${a * c + b}$`])), `L'image de ${c} par cette fonction vaut ${a * c + b}<br>On peut aussi dire que ${a * c + b} est l'image de ${c} par cette fonction`], this.spacing);
-					};
-					break;
-				case 2:
-					var j = 0; // pour la sous-numérotation
-					// les variables a,b,c changent sans refaire un appel à randint
-					texte = `Soit $f$ la fonction définie par l'expression algébrique $f(x)=$ ${a}$x+$${b}`;
-					if (sortie_html) {
-						// sous-question a/
-						texte += `<br>` + num_alpha(j) + ` Calculer l'image de ${c}`;
-						texte += `<br>`;
-						texte_corr = num_alpha(j) + ` Calculons l'image par $f$ de $x= ${c}$ :`;
-						texte_corr += `<br>$f(${mise_en_evidence('\\textit{\\textbf{x}}')})= ${a} ${mise_en_evidence('\\textit{\\textbf{x}}')}+${b}$`;
-						texte_corr += `<br>$f(${mise_en_evidence(c)})= ${a}\\times ${mise_en_evidence(c)}+${b}$`;
-						texte_corr += `<br>$f(${mise_en_evidence(c)})= ${a * c}+${b}$`;
-						texte_corr += `<br>$f(${mise_en_evidence(c)})= ${a * c + b}$`;
-						j++;
-						//sous question b/
-						texte += num_alpha(j) + ` Traduire ce calcul par une phrase contenant le mot image`;
-						texte_corr += `<br>` + num_alpha(j) + ` L'image de ${c} par la fonction $f$ vaut ${a * c + b}`;
-						texte_corr += `<br> On peut aussi dire que ${a * c + b} est l'image de ${c} par la fonction $f$`;
-					} else {
-						// sous-question a/ et b/
-						texte += tex_enumerate([`Calculer l'image de ${c}`, `Traduire ce calcul par une phrase contenant le mot image`], this.spacing);
-						texte_corr = tex_enumerate([`Calculons l'image par $f$ de $x= ${c}$ :
-							<br>$f(${mise_en_evidence('\\textit{\\textbf{x}}')})= ${a} ${mise_en_evidence('\\textit{\\textbf{x}}')}+${b}$
-							<br>$f(${mise_en_evidence(c)})= ${a}\\times ${mise_en_evidence(c)}+${b}$
-							<br>$f(${mise_en_evidence(c)})= ${a * c}+${b}$
-							<br>$f(${mise_en_evidence(c)})= ${a * c + b}$`, `L'image de ${c} par la fonction $f$ vaut ${a * c + b}
-							<br> On peut aussi dire que ${a * c + b} est l'image de ${c} par la fonction $f$`
-						], this.spacing);
-					};
-					break;
-				case 3:
-					var j = 0; // pour la sous-numérotation
-					// les variables a,b,c changent sans refaire un appel à randint
-					texte = `Soit $g$ la fonction définie par $g:x\\longmapsto$ ${a}$x+$${b}`;
-					if (sortie_html) {
-						// sous-question a/
-						texte += `<br>` + num_alpha(j) + ` Calculer l'image de ${c}`;
-						texte += `<br>`;
-						texte_corr = num_alpha(j) + ` Calculons l'image par $g$ de $x= ${c}$ :`;
-						texte_corr += `<br>$g:${mise_en_evidence('\\textit{\\textbf{x}}')}\\longmapsto ${a} ${mise_en_evidence('\\textit{\\textbf{x}}')}+${b}$`;
-						texte_corr += `<br>$g:${mise_en_evidence(c)}\\longmapsto ${a}\\times${mise_en_evidence(c)}+${b}$`;
-						texte_corr += `<br>$g:${mise_en_evidence(c)}\\longmapsto ${a * c}+${b}$`;
-						texte_corr += `<br>$g:${mise_en_evidence(c)}\\longmapsto ${a * c + b}$`;
-						j++;
-						//sous question b/
-						texte += num_alpha(j) + ` Traduire ce calcul par une phrase contenant le mot image`;
-						texte_corr += `<br>` + num_alpha(j) + ` L'image de ${c} par la fonction $g$ vaut ${a * c + b}`;
-						texte_corr += `<br> On peut aussi dire que ${a * c + b} est l'image de ${c} par la fonction $g$`;
-					} else {
-						// sous-question a/ et b/
-						texte += tex_enumerate([`Calculer l'image de ${c}`, `Traduire ce calcul par une phrase contenant le mot image`], this.spacing);
-						texte_corr = tex_enumerate([`Calculons l'image par $g$ de $x= ${c}$ :
-							<br>$g:${mise_en_evidence('\\textit{\\textbf{x}}')}\\longmapsto ${a} ${mise_en_evidence('\\textit{\\textbf{x}}')}+${b}$
-							<br>$g:${mise_en_evidence(c)}\\longmapsto ${a}\\times ${mise_en_evidence(c)}+${b}$
-							<br>$g:${mise_en_evidence(c)}\\longmapsto ${a * c}+${b}$
-							<br>$g:${mise_en_evidence(c)}\\longmapsto ${a * c + b}$`, `L'image de ${c} par la fonction $g$ vaut ${a * c + b}
-							<br> On peut aussi dire que ${a * c + b} est l'image de ${c} par la fonction $g$`
-						], this.spacing);
-					};
-					break;
-				case 4:
-					texte = ``;
-					texte_corr = ``;
-					texte_corr += `Calculer avec un diagramme `;
-					var j = 0; // pour la sous-numérotation
-					// les variables a,b,c changent sans refaire un appel à randint
-					texte += `Soit la fonction $h$ définie par le diagramme `;
-					if (sortie_html) {
-						// sous-question a/
-						texte += `<div id="${id_du_div}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-						SVG_machine_diag_3F12(id_du_div, 800, 100, 'h', 'x', [['' + a, a + 'x'], ['' + b, a + 'x+' + b]]);
-						texte += num_alpha(j) + ` Calculer l'image de ${c}`;
-						texte += `<br>`;
-						texte_corr += `<br>`;
-						texte_corr += num_alpha(j) + ` Calculons l'image par $h$ de $x=$ ${c} :`;
-						texte_corr += `<div id="${id_du_div_corr}" style="width: ${pourcentage}"; display : table "></div>`;
-						SVG_machine_diag_3F12(id_du_div_corr, 800, 100, 'h', '' + c, [['' + a, '' + (a * c)], ['' + b, '' + (a * c + b)]]);
-						j++;
-						//sous question b/
-						texte += num_alpha(j) + ` Traduire ce calcul par une phrase contenant le mot image`;
-						texte_corr += `<br>` + num_alpha(j) + ` L'image de ${c} par la fonction $h$ vaut ${a * c + b}`;
-						texte_corr += `<br> On peut aussi dire que ${a * c + b} est l'image de ${c} par la fonction $h$`;
-					} else {
-						texte += `<br>` + tikz_machine_diag(`h`, `x`, [[`\\times ` + a, a + `x`], [`+` + b, a + `x+` + b]]);
-						// sous-question a/ et b/
-						texte += tex_enumerate([`Calculer l'image de ${c}`, `Traduire ce calcul par une phrase contenant le mot image`], this.spacing);
-						texte_corr = tex_enumerate(
-							[`Calculons l'image par $g$ de $x=$ ${c} :<br>` + tikz_machine_diag(`h`, c, [[`\\times ` + a, (a * c)], [`+` + b, (a * c + b)]]),
-							`L'image de ${c} par la fonction $g$ vaut ${a * c + b}
-						 	<br> On peut aussi dire que ${a * c + b} est l'image de ${c} par la fonction $g$`
-							], this.spacing);
-					};
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Règle à travailler', 5, "1 : &Agrave; partir d'un programme de calcul\n2 : &Agrave; partir de l'expression algébrique sous forme f(x) = ...\n3 : &Agrave; partir de l'expression algébrique sous forme f : x --> ...\n4 : &Agrave; partir d'un diagramme\n5 : Mélange"];
-};
-
-/**
- * 3A10 - Division Euclidienne; diviseurs, multiples, critères de divisibilité
- * Exercice bilan
- * @author Sébastien Lozano
- */
-
-function DivisionEuclidienne_multiplesDiviseurs_Criteres() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.sup = 1;
-	this.titre = "Division Euclidienne - Diviseurs - Multiples";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = `Divisions euclidiennes - Diviseurs - Multiples.`;
-	//sortie_html ? this.spacing = 3 : this.spacing = 2;
-	sortie_html ? this.spacing = 1 : this.spacing = 2;
-	//sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 2;
-	this.nb_questions = 5;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//this.bouton_aide = '';
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheArithmetique-3A10.pdf", "Aide mémoire sur la division euclidienne (Sébastien Lozano)", "Aide mémoire")
-			//this.bouton_aide += modal_video('conteMathsNombresPremiers','videos/LesNombresPremiers.mp4','Petit conte mathématique','Intro Vidéo');
-		} else { // sortie LaTeX
-		};
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = [1, 2, 3, 4, 5];
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-
-		for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-			var dividende;
-			var diviseur;
-			var quotient;
-			var reste;
-
-			switch (type_de_questions) {
-				case 1: // plus grand reste dans une division euclidienne
-					diviseur = randint(2, 99);
-					texte = `Dire quel est le plus grand reste possible dans une division euclidienne par ${diviseur}.`;
-					texte_corr = `Si on divise par ${diviseur}, il ne peut pas rester plus de ${diviseur - 1}, sinon c'est qu'on peut encore ajouter au moins 1 fois ${diviseur} dans le dividende et donc 1 au quotient.`;
-					break;
-				case 2: // quotient et reste d'une division euclidienne donnée
-					diviseur = randint(2, 99);
-					dividende = randint(1001, 99999);
-					quotient = Math.trunc(dividende / diviseur);
-					reste = dividende % diviseur;
-
-					texte = `On a ${nombre_avec_espace(dividende)}=${nombre_avec_espace(diviseur)}$\\times$${nombre_avec_espace(quotient)} $+$ ${nombre_avec_espace(reste)}`;
-					texte += `<br>`;
-					texte += `Écrire le quotient et le reste de la division euclidienne de ${nombre_avec_espace(dividende)} par ${diviseur}.`;
-					texte_corr = `Dans la division euclidienne de ${nombre_avec_espace(dividende)} par ${diviseur}, le quotient vaut ${nombre_avec_espace(quotient)} et le reste ${reste}.`;
-					break;
-				case 3: // caractérisation des multiples et diviseurs par le reste de la division euclidienne
-					dividende = randint(101, 9999);
-					let rg_diviseur; // rang du diviseur choisi
-					if (liste_diviseurs(dividende).length % 2 == 0) {//si il y a un nombre pair de diviseurs on prend le (n/2+1) eme
-						rg_diviseur = liste_diviseurs(dividende).length / 2 + 1;
-					} else { // il y a nbre impair de diviseurs on prend le ((n-1)/2 +1) eme
-						rg_diviseur = (liste_diviseurs(dividende).length - 1) / 2 + 1;
-					}
-					diviseur = liste_diviseurs(dividende)[rg_diviseur - 1]; // on choisit le diviseur central de dividende, ATTENTION rang des tableaux commence à 0 
-					let candidats_diviseurs = [diviseur - 1, diviseur, diviseur + 1]; // on prend l'entier précédent et le successeur de ce diviseur
-					// Faut-il que je conditionne pour éviter le diviseur 1 ?
-					candidats_diviseurs = shuffle(candidats_diviseurs); // on mélange le tableau
-					texte = 'Les trois divisions euclidiennes suivantes sont exactes : <br>';
-					texte += `${nombre_avec_espace(dividende)} = ${nombre_avec_espace(candidats_diviseurs[0])}$\\times$${nombre_avec_espace(Math.trunc(dividende / candidats_diviseurs[0]))} $+$ ${nombre_avec_espace(dividende % candidats_diviseurs[0])}`;
-					texte += `<br>`;
-					texte += `${nombre_avec_espace(dividende)} = ${nombre_avec_espace(candidats_diviseurs[1])}$\\times$${nombre_avec_espace(Math.trunc(dividende / candidats_diviseurs[1]))} $+$ ${nombre_avec_espace(dividende % candidats_diviseurs[1])}`;
-					texte += `<br>`;
-					texte += `${nombre_avec_espace(dividende)} = ${nombre_avec_espace(candidats_diviseurs[2])}$\\times$${nombre_avec_espace(Math.trunc(dividende / candidats_diviseurs[2]))} $+$ ${nombre_avec_espace(dividende % candidats_diviseurs[2])}`;
-					texte += `<br>`;
-					texte += `Sans calculer, dire si les nombres ${nombre_avec_espace(candidats_diviseurs[0])}; ${nombre_avec_espace(candidats_diviseurs[1])}; ${nombre_avec_espace(candidats_diviseurs[2])} sont des diviseurs de ${nombre_avec_espace(dividende)}. Justifier.`;
-					texte_corr = ``;
-					if (egal(dividende % candidats_diviseurs[0], 0)) { //egal() est une fonction de JC pour éviter les problèmes de virgule flottante
-						texte_corr += `Le reste de la division euclienne de ${nombre_avec_espace(dividende)} par ${nombre_avec_espace(candidats_diviseurs[0])} vaut 0 donc ${nombre_avec_espace(candidats_diviseurs[0])} est un diviseur de ${nombre_avec_espace(dividende)}`;
-					} else {
-						texte_corr += `Le reste de la division euclienne de ${nombre_avec_espace(dividende)} par ${nombre_avec_espace(candidats_diviseurs[0])} ne vaut pas 0 donc ${nombre_avec_espace(candidats_diviseurs[0])} n'est pas un diviseur de ${nombre_avec_espace(dividende)}`;
-					}
-					texte_corr += `<br>`;
-					if (egal(dividende % candidats_diviseurs[1], 0)) { //egal() est une fonction de JC pour éviter les problèmes de virgule flottante
-						texte_corr += `Le reste de la division euclienne de ${nombre_avec_espace(dividende)} par ${nombre_avec_espace(candidats_diviseurs[1])} vaut 0 donc ${nombre_avec_espace(candidats_diviseurs[1])} divise ${nombre_avec_espace(dividende)}`;
-					} else {
-						texte_corr += `Le reste de la division euclienne de ${nombre_avec_espace(dividende)} par ${nombre_avec_espace(candidats_diviseurs[1])} ne vaut pas 0 donc ${nombre_avec_espace(candidats_diviseurs[1])} ne divise pas ${nombre_avec_espace(dividende)}`;
-					}
-					texte_corr += `<br>`;
-					if (egal(dividende % candidats_diviseurs[2], 0)) { //egal() est une fonction de JC pour éviter les problèmes de virgule flottante
-						texte_corr += `Le reste de la division euclienne de ${nombre_avec_espace(dividende)} par ${nombre_avec_espace(candidats_diviseurs[2])} vaut 0 donc ${nombre_avec_espace(dividende)} est divisible par ${nombre_avec_espace(candidats_diviseurs[2])}`;
-					} else {
-						texte_corr += `Le reste de la division euclienne de ${nombre_avec_espace(dividende)} par ${nombre_avec_espace(candidats_diviseurs[2])} ne vaut pas 0 donc ${nombre_avec_espace(dividende)} n'est pas divisible par ${nombre_avec_espace(candidats_diviseurs[2])}`;
-					}
-					texte_corr += `<br>`;
-					break;
-				case 4: // vocabulaire diviseurs et multiples
-					// on déclare des tableaux utiles 
-					let diviseurs = [];
-					let multiplicateurs = [];
-					let multiples = [];
-					let quotients = [];
-					let textes = [];
-					let textes_corr = [];
-					// on tire au hasard 4 diviseurs différents entre 2 et 999 et 4 multiplicateurs différents entre 2 et 9 
-					diviseurs = [randint(2, 999), randint(2, 999, [diviseurs[0]]), randint(2, 999, [diviseurs[0], diviseurs[1]]), randint(2, 999, [diviseurs[0], diviseurs[1], diviseurs[2]])];
-					multiplicateurs = [randint(2, 9), randint(2, 9, [multiplicateurs[0]]), randint(2, 9, [multiplicateurs[0], multiplicateurs[1]]), randint(2, 9, [multiplicateurs[0], multiplicateurs[1], multiplicateurs[2]])];
-					// on calcule les multiples
-					for (let j = 0; j < 4; j++) {
-						multiples[j] = diviseurs[j] * multiplicateurs[j];
-						quotients[j] = multiples[j] / diviseurs[j];
-						diviseurs[j] = nombre_avec_espace(diviseurs[j]);
-						multiples[j] = nombre_avec_espace(multiples[j]);
-						quotients[j] = nombre_avec_espace(quotients[j]);
-					};
-					// on crée les phrases 
-					textes[0] = `${diviseurs[0]} $\\ldots\\ldots\\ldots\\ldots$ ${multiples[0]}`;
-					textes_corr[0] = `${diviseurs[0]} est un diviseur de ${multiples[0]} car ${multiples[0]}=${diviseurs[0]}$\\times$${quotients[0]}`;
-					textes[1] = `${diviseurs[1]} $\\ldots\\ldots\\ldots\\ldots$ ${multiples[1]}`;
-					textes_corr[1] = `${diviseurs[1]} est un diviseur de ${multiples[1]} car ${multiples[1]}=${diviseurs[1]}$\\times$${quotients[1]}`;
-					textes[2] = `${multiples[2]} $\\ldots\\ldots\\ldots\\ldots$ ${diviseurs[2]}`;
-					textes_corr[2] = `${multiples[2]} est un multiple de ${diviseurs[2]} car ${multiples[2]}=${diviseurs[2]}$\\times$${quotients[2]}`;
-					textes[3] = `${multiples[3]} $\\ldots\\ldots\\ldots\\ldots$ ${diviseurs[3]}`;
-					textes_corr[3] = `${multiples[3]} est un multiple de ${diviseurs[3]} car ${multiples[3]}=${diviseurs[3]}$\\times$${quotients[3]}`;
-					// on ajoute deux cas ni multiple ni diviseur
-					// on choisit deux nombres
-					let n1 = nombre_avec_espace(randint(2, 999, [diviseurs[0], diviseurs[1], diviseurs[2], diviseurs[3]]));
-					let p1 = nombre_avec_espace(randint(2, 999, [diviseurs[0], diviseurs[1], diviseurs[2], diviseurs[3], n1]));
-					// on choisit un autre qui n'est pas dans la liste des diviseurs de n1
-					let n2 = nombre_avec_espace(randint(2, 999, liste_diviseurs(n1)));
-					let p2 = nombre_avec_espace(randint(2, 999, liste_diviseurs(p1)));
-					textes[4] = `${n1} $\\ldots\\ldots\\ldots\\ldots$ ${n2}`;
-					textes_corr[4] = `${n1} n'est ni un multiple ni un diviseur de ${n2} car ${n1}=${n2}$\\times$${Math.trunc(n1 / n2)}+${texte_en_couleur(n1 % n2)} et ${n2}=${n1}$\\times$${Math.trunc(n2 / n1)}+${texte_en_couleur(n2 % n1)}`;
-					textes[5] = `${p2} $\\ldots\\ldots\\ldots\\ldots$ ${p1}`;
-					textes_corr[5] = `${p2} n'est ni un multiple ni un diviseur de ${p1} car ${p1}=${p2}$\\times$${Math.trunc(p1 / p2)}+${texte_en_couleur(p1 % p2)} et ${p2}=${p1}$\\times$${Math.trunc(p2 / p1)}+${texte_en_couleur(p2 % p1)}`;
-					// on mélange pour que l'ordre change!
-					shuffle2tableaux(textes, textes_corr);
-					texte = `Avec la calculatrice, compléter chaque phrase avec "est un diviseur de" ou "est un multiple de" ou "n'est ni un diviseur ni un multiple de".`;
-					texte += `<br>`;
-					texte_corr = ``;
-					for (let j = 0; j < 5; j++) {
-						texte += textes[j];
-						texte += `<br>`;
-						texte_corr += textes_corr[j];
-						texte_corr += `<br>`;
-					};
-					texte += textes[5];
-					//texte +=`<br>`;
-					texte_corr += textes_corr[5];
-					texte_corr += `<br>`;
-					break;
-				case 5: // liste des diviseurs
-					// on définit un tableau pour les choix du nombre dont on veut les diviseurs
-					// 3 parmis 2,99 y compris les premiers et 1 parmis les entiers à 3 chiffres ayant au moins 8 diviseurs, il y en a 223 !
-					let tableau_de_choix = [];
-					tableau_de_choix = [randint(2, 99), randint(2, 99, [tableau_de_choix[0]]), randint(2, 99, [tableau_de_choix[0], tableau_de_choix[1]]), randint(2, 99, [tableau_de_choix[0], tableau_de_choix[1], tableau_de_choix[2]])];
-					let tableau_de_choix_3chiffres = [];
-					for (let m = 101; m < 999; m++) {
-						if (liste_diviseurs(m).length > 8) {
-							tableau_de_choix_3chiffres.push(m);
-						};
-					};
-					// on ajoute un nombre à trois chiffre avec au moins 8 diviseurs dans les choix possibles
-					let rg_Nb_3chiffres = randint(0, (tableau_de_choix_3chiffres.length - 1));
-					tableau_de_choix.push(tableau_de_choix_3chiffres[rg_Nb_3chiffres]);
-					let N; // on déclare le nombre dont on va chercher les diviseurs
-					let rg_N; // pour tirer le rang du nombre dans le tableau des choix
-					rg_N = randint(0, (tableau_de_choix.length - 1));
-					N = tableau_de_choix[rg_N];
-					texte = `Écrire la liste de tous les diviseurs de ${N}.`;
-					texte_corr = `Pour trouver la liste des diviseurs de ${N} on cherche tous les produits de deux facteurs qui donnent ${N}. En écrivant toujours le plus petit facteur en premier.<br>`;
-					texte_corr += `Il est suffisant de chercher des diviseurs inférieurs au plus grand nombre dont le carré vaut ${N}, par exemple ici, ${Math.trunc(Math.sqrt(N))}$\\times$${Math.trunc(Math.sqrt(N))} = ${Math.trunc(Math.sqrt(N)) * Math.trunc(Math.sqrt(N))}<${N}`;
-					texte_corr += ` et ${Math.trunc(Math.sqrt(N)) + 1}$\\times$${Math.trunc(Math.sqrt(N)) + 1} = ${(Math.trunc(Math.sqrt(N)) + 1) * (Math.trunc(Math.sqrt(N)) + 1)}>${N} donc il suffit d'arrêter la recherche de facteur à ${Math.trunc(Math.sqrt(N))}.`;
-					texte_corr += ` En effet, si ${N} est le produit de deux entiers p$\\times$q avec p < q alors si p$\\times$p > ${N} c'est que q$\\times$q < ${N} mais dans ce cas p serait supérieur à q sinon p$\\times$q serait inférieur à ${N} ce qui ne doit pas être le cas.<br>`
-					if (liste_diviseurs(N).length % 2 == 0) {//si il y a un nombre pair de diviseurs
-						for (let m = 0; m < (liste_diviseurs(N).length / 2); m++) {
-							texte_corr += `` + liste_diviseurs(N)[m] + `$\\times$` + liste_diviseurs(N)[(liste_diviseurs(N).length - m - 1)] + ` = ${N}<br>`;
-						};
-					} else {
-						for (let m = 0; m < ((liste_diviseurs(N).length - 1) / 2); m++) {
-							texte_corr += `` + liste_diviseurs(N)[m] + `$\\times$` + liste_diviseurs(N)[(liste_diviseurs(N).length - m - 1)] + `<br>`;
-						};
-						texte_corr += `` + liste_diviseurs(N)[(liste_diviseurs(N).length - 1) / 2] + `$\\times$` + liste_diviseurs(N)[(liste_diviseurs(N).length - 1) / 2] + ` = ${N}<br>`;
-					};
-					texte_corr += `Chacun des facteurs de la liste ci-dessus est un diviseur de ${N}.<br>`;
-					texte_corr += `La liste des diviseurs de ${N} est donc `;
-					texte_corr += `1`;
-					for (let w = 1; w < liste_diviseurs(N).length; w++) {
-						texte_corr += ` ; ` + liste_diviseurs(N)[w];
-					};
-					texte_corr += `.`;
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	//this.besoin_formulaire_numerique = ['Règle à travailler',5,"1 : Produit de deux puissances de même base\n2 : Quotient de deux puissances de même base\n3 : Puissance de puissance\n4 : Produit de puissances de même exposant\n5 : Mélange"]; 
-};
-
-/**
- * 3A11 justifier la non primalité réinvestissement des critères de divisibilité
- * Nombres à 3 ou 4 chiffres, un multiple de 2, de 3, de 5, de 7, de 11, sous forme d'un produit de deux nombres premiers inférieurs à 100
- * et un nombre premier inferieur à 529
- * dans cet exo on n'utilise pas les critères par 7 et 11
- * @author Sébastien Lozano
- */
-function Premier_ou_pas() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Primalité ou pas";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = `Justifier que les nombres suivants sont premiers ou pas. Penser aux critères de divisibilité.`;
-	//sortie_html ? this.spacing = 3 : this.spacing = 2;
-	sortie_html ? this.spacing = 1 : this.spacing = 2;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-
-
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 2;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-	this.nb_questions_modifiable = false;
-	// console.log(Number(this.sup)==1);
-	// if (Number(this.sup)==1) {
-	// 	this.nb_questions = 4;
-	// } else {
-	// 	this.nb_questions = 5;
-	// }
-	this.liste_packages = `bclogo`;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//this.bouton_aide = '';
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheArithmetique-3A11.pdf", "Aide mémoire sur les nombres premiers (Sébastien Lozano)", "Aide mémoire")
-			this.bouton_aide += modal_video('conteMathsNombresPremiers', 'videos/LesNombresPremiers.mp4', 'Petit conte mathématique - Les Nombres Premiers', 'Intro Vidéo');
-		} else { // sortie LaTeX
-		};
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-
-		let type_de_questions_disponibles;// = [1, 2, 3, 6, 7];
-		if (Number(this.sup) == 1) {
-			this.nb_questions = 4;
-			type_de_questions_disponibles = [1, 2, 3, 8];
-		} else {
-			this.nb_questions = 5;
-			type_de_questions_disponibles = [1, 2, 3, 6, 7];
-		}
-		type_de_questions_disponibles = shuffle(type_de_questions_disponibles); // on mélange l'ordre des questions
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-
-		let string_rappel = `Cette liste des nombres premiers inférieurs à 100 pourra être utile : <br>` + crible_eratosthene_n(100)[0];
-		for (let k = 1; k < crible_eratosthene_n(100).length; k++) {
-			string_rappel += `, ` + crible_eratosthene_n(100)[k];
-		};
-		string_rappel += `.`;
-
-		this.introduction = warn_message(string_rappel, `nombres`, `Coup de pouce`);
-
-		for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-
-			type_de_questions = liste_type_de_questions[i];
-
-			var N; // le nombre de la question
-			let r;
-			let tab_premiers_a_tester;
-
-			switch (type_de_questions) {
-				case 1: // nombre pair
-					N = 2 * randint(51, 4999);
-					texte = nombre_avec_espace(N);
-					texte_corr = `Comme ${nombre_avec_espace(N)} est pair, il admet donc au moins trois diviseurs qui sont 1, 2 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 2: // Multiple de 3
-					let sum = 0; // pour la valeur de la somme;
-					N = 3 * randint(34, 3333); // on initialise avant la boucle car on a peut être de la chance
-					while ((N % 2 == 0) || (N % 5 == 0)) {
-						N = 3 * randint(34, 3333);
-					};
-					texte = nombre_avec_espace(N);
-					texte_corr = `Comme ` + N.toString().charAt(0);
-					sum = Number(N.toString().charAt(0));
-					for (let k = 1; k < N.toString().length; k++) {
-						texte_corr += ` + ` + N.toString().charAt(k);
-						sum += Number(N.toString().charAt(k));
-					};
-					texte_corr += ` = ${sum} est un multiple de 3 donc ${nombre_avec_espace(N)} aussi, il admet donc au moins trois diviseurs qui sont 1, 3 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 3: // Multiple de 5
-					N = 5 * randint(20, 1999);
-					texte = nombre_avec_espace(N);
-					texte_corr = `Comme le dernier chiffre de ${nombre_avec_espace(N)} est un ${N.toString().charAt(N.toString().length - 1)} alors ${nombre_avec_espace(N)} est divisible par 5, `;
-					texte_corr += `il admet donc au moins trois diviseurs qui sont 1, 5 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 4: // Multiple de 7
-					let N_longueur; // pour la taille du string N
-					let N1; // pour la repetiton du critère
-					let N1_longueur; // pour la taille du string N1
-					let sum1; // pour la somme de la répétition du critère
-					N = 7 * randint(15, 1428);
-					texte = nombre_avec_espace(N);
-					N_longueur = N.toString().length;
-					texte_corr = ` 7 divise ${nombre_avec_espace(N)}, en effet : `;
-					texte_corr += `<br>`;
-					N1 = N;
-					N1_longueur = N_longueur;
-					sum1 = Number(N1.toString().substring(0, N1_longueur - 1)) + 5 * Number(N1.toString().charAt(N1_longueur - 1));
-					while (sum1 >= 56) {
-						texte_corr += `${N1.toString().substring(0, N1_longueur - 1)} + 5$\\times$${N1.toString().charAt(N1_longueur - 1)}`;
-						texte_corr += ` = ${Number(N1.toString().substring(0, N1_longueur - 1)) + 5 * Number(N1.toString().charAt(N1_longueur - 1))}`;
-						texte_corr += `<br>`;
-						N1 = sum1;
-						N1_longueur = N1.toString().length;
-						sum1 = Number(N1.toString().substring(0, N1_longueur - 1)) + 5 * Number(N1.toString().charAt(N1_longueur - 1));
-					};
-					texte_corr += `Comme ${N1.toString().substring(0, N1_longueur - 1)} + 5$\\times$${N1.toString().charAt(N1_longueur - 1)} = ${sum1} est un multiple de 7 alors 7 divise ${N} aussi `;
-					texte_corr += `qui admet donc au moins trois diviseurs : 1, 7 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 5: // multiple de 11
-					let even_sum; // pour la somme des chiffres de rang impair
-					let odd_sum; // pour la somme des chiffres de rang pair
-					N = 11 * randint(10, 909);
-					texte = nombre_avec_espace(N);
-					texte_corr = `D'une part, la somme des chiffres de rang impair de ${nombre_avec_espace(N)} vaut `;
-					if (Number(N.toString().length) % 2 == 0) { // si N a un nombre pair de chiffres
-						even_sum = Number(N.toString().charAt(1));
-						texte_corr += N.toString().charAt(1);
-						for (let k = 3; k < N.toString().length; k++) {
-							if (k % 2 == 1) {
-								texte_corr += ` + ` + N.toString().charAt(k);
-								even_sum += Number(N.toString().charAt(k));
-							};
-						};
-						texte_corr += ` = ` + even_sum + ` <br> `;
-					} else { // sinon N a un nombre impair de chiffres
-						even_sum = Number(N.toString().charAt(0));
-						texte_corr += N.toString().charAt(0);
-						for (let m = 1; m < N.toString().length; m++) {
-							if (m % 2 == 0) {
-								texte_corr += ` + ` + N.toString().charAt(m);
-								even_sum += Number(N.toString().charAt(m));
-							};
-
-						};
-						texte_corr += ` = ` + even_sum + `<br> `;
-					};
-					texte_corr += `d'autre part, la somme des chiffres de rang pair de ${nombre_avec_espace(N)} vaut `;
-					if (Number(N.toString().length) % 2 == 0) { // si N a un nombre pair de chiffres
-						odd_sum = Number(N.toString().charAt(0));
-						texte_corr += N.toString().charAt(0);
-						for (let k = 1; k < N.toString().length; k++) {
-							if (k % 2 == 0) {
-								texte_corr += ` + ` + N.toString().charAt(k);
-								odd_sum += Number(N.toString().charAt(k));
-							};
-						};
-						texte_corr += ` = ` + odd_sum + ` <br> `;
-					} else { // sinon N a un nombre impair de chiffres
-						odd_sum = Number(N.toString().charAt(1));
-						texte_corr += N.toString().charAt(1);
-						for (let m = 3; m < N.toString().length; m++) {
-							if (m % 2 == 1) {
-								texte_corr += ` + ` + N.toString().charAt(m);
-								odd_sum += Number(N.toString().charAt(m));
-							};
-
-						};
-						texte_corr += ` = ` + odd_sum + `<br> `;
-					};
-					texte_corr += `la différence entre la somme des chiffres de rangs pairs et celle des chiffres de rangs impairs vaut `;
-					if ((odd_sum - even_sum) == 0) {
-						texte_corr += `${odd_sum - even_sum}, `;
-
-					} else {
-						texte_corr += `${Math.abs(odd_sum - even_sum)} qui est un multiple de 11, `;
-					};
-					texte_corr += `<br>`;
-					texte_corr += ` cela signifie que ${nombre_avec_espace(N)} est divisible par 11, il admet donc au moins trois diviseurs qui sont 1, 11 et lui-même,`;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 6: // produit de deux nombres premiers inférieurs à 100
-					// rang du premier facteur premier
-					let r1 = randint(0, crible_eratosthene_n(100).length - 1);
-					// rang du second facteur premier
-					let r2 = randint(0, crible_eratosthene_n(100).length - 1);
-					let prime1 = crible_eratosthene_n(100)[r1]; // on tire un nombre premier inférieur à 100, il n'y en a que 25!
-					let prime2 = crible_eratosthene_n(100)[r2]; // on tire un autre nombre premier inférieur à 100, ça peut être le même qu'avant!
-					N = prime1 + `$\\times$` + prime2;
-					texte = N;
-					texte_corr = `${N} est le produit de ${prime1} et de ${prime2}, il admet donc au moins `;
-					if (prime1 == prime2) {
-						texte_corr += `trois divisieurs qui sont 1, ${prime1} et lui-même ${N}=${nombre_avec_espace(prime1 * prime2)}, `;
-					} else {
-						texte_corr += `quatre diviseurs qui sont 1, ${prime1}, ${prime2} et lui-même ${N}=${nombre_avec_espace(prime1 * prime2)}, `;
-					};
-					texte_corr += texte_en_couleur_et_gras(`${N} = ` + nombre_avec_espace(prime1 * prime2) + ` n'est donc pas premier.`);
-					break;
-				case 7: // nombre premier inférieur à 529
-					// rang du nombre premier choisi
-					r = randint(0, crible_eratosthene_n(529).length - 1);
-					N = crible_eratosthene_n(529)[r]; //on choisit un nombre premier inférieur à 529
-					texte = N + ``;
-					tab_premiers_a_tester = crible_eratosthene_n(Math.trunc(Math.sqrt(N)));
-					//texte_corr = `Testons la divisibilité de ${N} par tous les nombres premiers inférieurs à $\\sqrt{${N}}$, c'est à dire par les nombres `;
-					texte_corr = `En effectuant la division euclidienne de ${N} par tous les nombres premiers inférieurs à $\\sqrt{${N}}$, c'est à dire par les nombres `;
-					texte_corr += tab_premiers_a_tester[0];
-					for (let k = 1; k < tab_premiers_a_tester.length; k++) {
-						texte_corr += `, ` + tab_premiers_a_tester[k];
-					};
-					//texte_corr += `.`;
-					// texte_corr += `<br> Aucun de ces nombres premiers ne divise ${N}, `;
-					texte_corr += `, le reste n'est jamais nul.`;
-					// texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` est donc un nombre premier.`);
-					texte_corr += `<br>` + texte_en_couleur_et_gras(nombre_avec_espace(N) + ` est donc un nombre premier.`);
-					break;
-				case 8: // nombre premier inférieur à 100 pour permettre les tests de divisibilité sans calculatrice
-					// rang du nombre premier choisi
-					r = randint(0, crible_eratosthene_n(100).length - 1);
-					N = crible_eratosthene_n(100)[r]; //on choisit un nombre premier inférieur à 529
-					texte = N + ``;
-					tab_premiers_a_tester = crible_eratosthene_n(Math.trunc(Math.sqrt(N)));
-					//texte_corr = `Testons la divisibilité de ${N} par tous les nombres premiers inférieurs à $\\sqrt{${N}}$, c'est à dire par les nombres `;
-					texte_corr = `En effectuant la division euclidienne de ${N} par tous les nombres premiers inférieurs à $\\sqrt{${N}}$, c'est à dire par les nombres `;
-					texte_corr += tab_premiers_a_tester[0];
-					for (let k = 1; k < tab_premiers_a_tester.length; k++) {
-						texte_corr += `, ` + tab_premiers_a_tester[k];
-					};
-					//texte_corr += `.`;
-					// texte_corr += `<br> Aucun de ces nombres premiers ne divise ${N}, `;
-					texte_corr += `, le reste n'est jamais nul.`;
-					// texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` est donc un nombre premier.`);
-					texte_corr += `<br>` + texte_en_couleur_et_gras(nombre_avec_espace(N) + ` est donc un nombre premier.`);
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Difficulté', 2, "1 : Sans Calculatrice\n2 : Avec calculatrice"];
-};
-
-/**
- * 3A11-1 justifier la non primalité réinvestissement des critères de divisibilité
- * Nombres à 3 ou 4 chiffres, un multiple de 2, de 3, de 5, de 7, de 11, sous forme d'un produit de deux nombres premiers inférieurs à 100
- * et un nombre premier inferieur à 529
- * variante de 3A11 avec les critères par 7 et 11 en plus
- * @author Sébastien Lozano
- */
-function Premier_ou_pas_critere_par7_par11() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Primalité ou pas - Variante avec les critères de divisibilité par 7 et par 11";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = `Justifier que les nombres suivants sont premiers ou pas. Penser aux critères de divisibilité.`;
-	//this.consigne += `<br>`;
-	sortie_html ? this.spacing = 3 : this.spacing = 2;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	this.nb_questions = 7;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 2;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-	this.liste_packages = `bclogo`;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//this.bouton_aide = '';
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheArithmetique-3A11.pdf", "Aide mémoire sur les nombres premiers (Sébastien Lozano)", "Aide mémoire")
-			this.bouton_aide += modal_video('conteMathsNombresPremiers', 'videos/LesNombresPremiers.mp4', 'Petit conte mathématique - Les Nombres Premiers', 'Intro Vidéo');
-		} else { // sortie LaTeX
-		};
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = [1, 2, 3, 4, 5, 6, 7];
-		type_de_questions_disponibles = shuffle(type_de_questions_disponibles); // on mélange l'ordre des questions
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-
-		let string_rappel_b = `Ces critères de divisibilité pourront être utiles :`;
-		if (sortie_html) {
-			string_rappel_b += `<br>`;
-			string_rappel_b += `- Un nombre est divisible par 7 si la somme de son nombre de dizaines et de cinq fois son chiffre des unités l’est.<br>`;
-			string_rappel_b += `- Un nombre est divisible par 11 si la différence entre la somme de ses chiffres de rangs pairs et la somme de ses chiffres de rangs impairs est nulle ou égale à un multiple de 11.`;
-			string_rappel_b += `<br> <br>`;
-		} else {
-			string_rappel_b += itemize([
-				`Un nombre est divisible par 7 si la somme de son nombre de dizaines et de cinq fois son chiffre des unités l’est.`,
-				`Un nombre est divisible par 11 si la différence entre la somme de ses chiffres de rangs pairs et la somme de ses chiffres de rangs impairs est nulle ou égale à un multiple de 11.`
-			]);
-			string_rappel_b += `\\par\\vspace{0.5cm}`;
-		};
-		string_rappel_b += `Ainsi que cette liste des nombres premiers inférieurs à 100 : `;
-		if (sortie_html) {
-			string_rappel_b += `<br>`;
-		} else {
-			string_rappel_b += `\\par\\vspace{0.25cm}`;
-		};
-		string_rappel_b += crible_eratosthene_n(100)[0];
-		for (let k = 1; k < crible_eratosthene_n(100).length; k++) {
-			string_rappel_b += `, ` + crible_eratosthene_n(100)[k];
-		};
-		string_rappel_b += `.`;
-
-		this.introduction = warn_message(string_rappel_b, `nombres`, `Coup de pouce`);
-
-		for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-			var N; // le nombre de la question
-
-			switch (type_de_questions) {
-				case 1: // nombre pair
-					N = 2 * randint(51, 4999);
-					texte = nombre_avec_espace(N);
-					texte_corr = `Comme ${nombre_avec_espace(N)} est pair, il admet donc au moins trois diviseurs qui sont 1, 2 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 2: // Multiple de 3
-					let sum = 0; // pour la valeur de la somme;
-					N = 3 * randint(34, 3333);// on initialise avant la boucle car on a peut être de la chance
-					while ((N % 2 == 0) || (N % 5 == 0)) {
-						N = 3 * randint(34, 3333);
-					};
-					texte = nombre_avec_espace(N);
-					texte_corr = `Comme ` + N.toString().charAt(0);
-					sum = Number(N.toString().charAt(0));
-					for (let k = 1; k < N.toString().length; k++) {
-						texte_corr += ` + ` + N.toString().charAt(k);
-						sum += Number(N.toString().charAt(k));
-					};
-					texte_corr += ` = ${sum} est un multiple de 3 donc ${nombre_avec_espace(N)} aussi, il admet donc au moins trois diviseurs qui sont 1, 3 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 3: // Multiple de 5
-					N = 5 * randint(20, 1999);
-					texte = nombre_avec_espace(N);
-					texte_corr = `Comme le dernier chiffre de ${nombre_avec_espace(N)} est un ${N.toString().charAt(N.toString().length - 1)} alors ${nombre_avec_espace(N)} est divisible par 5, `;
-					texte_corr += `il admet donc au moins trois diviseurs qui sont 1, 5 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 4: // Multiple de 7
-					let N_longueur; // pour la taille du string N
-					let N1; // pour la repetiton du critère
-					let N1_longueur; // pour la taille du string N1
-					let sum1; // pour la somme de la répétition du critère
-					N = 7 * randint(15, 1428);
-					texte = nombre_avec_espace(N);
-					N_longueur = N.toString().length;
-					texte_corr = ` 7 divise ${nombre_avec_espace(N)}, en effet : `;
-					texte_corr += `<br>`;
-					N1 = N;
-					N1_longueur = N_longueur;
-					sum1 = Number(N1.toString().substring(0, N1_longueur - 1)) + 5 * Number(N1.toString().charAt(N1_longueur - 1));
-					while (sum1 >= 56) {
-						texte_corr += `${N1.toString().substring(0, N1_longueur - 1)} + 5$\\times$${N1.toString().charAt(N1_longueur - 1)}`;
-						texte_corr += ` = ${Number(N1.toString().substring(0, N1_longueur - 1)) + 5 * Number(N1.toString().charAt(N1_longueur - 1))}`;
-						texte_corr += `<br>`;
-						N1 = sum1;
-						N1_longueur = N1.toString().length;
-						sum1 = Number(N1.toString().substring(0, N1_longueur - 1)) + 5 * Number(N1.toString().charAt(N1_longueur - 1));
-					};
-					texte_corr += `Comme ${N1.toString().substring(0, N1_longueur - 1)} + 5$\\times$${N1.toString().charAt(N1_longueur - 1)} = ${sum1} est un multiple de 7 alors 7 divise ${N} aussi `;
-					texte_corr += `qui admet donc au moins trois diviseurs : 1, 7 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 5: // multiple de 11
-					let even_sum; // pour la somme des chiffres de rang impair
-					let odd_sum; // pour la somme des chiffres de rang pair
-					N = 11 * randint(10, 909);
-					texte = nombre_avec_espace(N);
-					texte_corr = `D'une part, la somme des chiffres de rang impair de ${nombre_avec_espace(N)} vaut `;
-					if (Number(N.toString().length) % 2 == 0) { // si N a un nombre pair de chiffres
-						even_sum = Number(N.toString().charAt(1));
-						texte_corr += N.toString().charAt(1);
-						for (let k = 3; k < N.toString().length; k++) {
-							if (k % 2 == 1) {
-								texte_corr += ` + ` + N.toString().charAt(k);
-								even_sum += Number(N.toString().charAt(k));
-							};
-						};
-						texte_corr += ` = ` + even_sum + ` <br> `;
-					} else { // sinon N a un nombre impair de chiffres
-						even_sum = Number(N.toString().charAt(0));
-						texte_corr += N.toString().charAt(0);
-						for (let m = 1; m < N.toString().length; m++) {
-							if (m % 2 == 0) {
-								texte_corr += ` + ` + N.toString().charAt(m);
-								even_sum += Number(N.toString().charAt(m));
-							};
-
-						};
-						texte_corr += ` = ` + even_sum + `<br> `;
-					};
-					texte_corr += `d'autre part, la somme des chiffres de rang pair de ${nombre_avec_espace(N)} vaut `;
-					if (Number(N.toString().length) % 2 == 0) { // si N a un nombre pair de chiffres
-						odd_sum = Number(N.toString().charAt(0));
-						texte_corr += N.toString().charAt(0);
-						for (let k = 1; k < N.toString().length; k++) {
-							if (k % 2 == 0) {
-								texte_corr += ` + ` + N.toString().charAt(k);
-								odd_sum += Number(N.toString().charAt(k));
-							};
-						};
-						texte_corr += ` = ` + odd_sum + ` <br> `;
-					} else { // sinon N a un nombre impair de chiffres
-						odd_sum = Number(N.toString().charAt(1));
-						texte_corr += N.toString().charAt(1);
-						for (let m = 3; m < N.toString().length; m++) {
-							if (m % 2 == 1) {
-								texte_corr += ` + ` + N.toString().charAt(m);
-								odd_sum += Number(N.toString().charAt(m));
-							};
-
-						};
-						texte_corr += ` = ` + odd_sum + `<br> `;
-					};
-					texte_corr += `la différence entre la somme des chiffres de rangs pairs et celle des chiffres de rangs impairs vaut `;
-					if ((odd_sum - even_sum) == 0) {
-						texte_corr += `${odd_sum - even_sum}, `;
-
-					} else {
-						texte_corr += `${Math.abs(odd_sum - even_sum)} qui est un multiple de 11, `;
-					};
-					texte_corr += `<br>`;
-					texte_corr += ` cela signifie que ${nombre_avec_espace(N)} est divisible par 11, il admet donc au moins trois diviseurs qui sont 1, 11 et lui-même, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(N) + ` n'est donc pas premier.`);
-					break;
-				case 6: // produit de deux nombres premiers inférieurs à 100
-					// rang du premier facteur premier
-					let r1 = randint(0, crible_eratosthene_n(100).length - 1);
-					// rang du second facteur premier
-					let r2 = randint(0, crible_eratosthene_n(100).length - 1);
-					let prime1 = crible_eratosthene_n(100)[r1]; // on tire un nombre premier inférieur à 100, il n'y en a que 25!
-					let prime2 = crible_eratosthene_n(100)[r2]; // on tire un autre nombre premier inférieur à 100, ça peut être le même qu'avant!
-					N = prime1 + `$\\times$` + prime2;
-					texte = N;
-					texte_corr = `${N} est le produit de ${prime1} et de ${prime2}, il admet donc au moins `;
-					if (prime1 == prime2) {
-						texte_corr += `trois divisieurs qui sont 1, ${prime1} et lui-même ${N}=${nombre_avec_espace(prime1 * prime2)}, `;
-					} else {
-						texte_corr += `quatre diviseurs qui sont 1, ${prime1}, ${prime2} et lui-même ${N}=${nombre_avec_espace(prime1 * prime2)}, `;
-					};
-					texte_corr += texte_en_couleur_et_gras(`${N} = ` + nombre_avec_espace(prime1 * prime2) + ` n'est donc pas premier.`);
-					break;
-				case 7: // nombre premier inférieur à 529
-					// rang du nombre premier choisi
-					let r = randint(0, crible_eratosthene_n(529).length - 1);
-					N = crible_eratosthene_n(529)[r]; //on choisit un nombre premier inférieur à 529
-					texte = N + ``;;
-					let tab_premiers_a_tester = crible_eratosthene_n(Math.trunc(Math.sqrt(N)));
-					texte_corr = `Testons la divisibilité de ${N} par tous les nombres premiers inférieurs à $\\sqrt{${N}}$, c'est à dire par les nombres `;
-					texte_corr += tab_premiers_a_tester[0];
-					for (let k = 1; k < tab_premiers_a_tester.length; k++) {
-						texte_corr += `, ` + tab_premiers_a_tester[k];
-					};
-					texte_corr += `.`;
-					texte_corr += `<br> Aucun de ces nombres premiers ne divise ${N}, `;
-					texte_corr += texte_en_couleur_et_gras(nombre_avec_espace(Number(N)) + ` n'est donc pas premier.`);
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	//this.besoin_formulaire_numerique = ['Règle à travailler',5,"1 : Produit de deux puissances de même base\n2 : Quotient de deux puissances de même base\n3 : Puissance de puissance\n4 : Produit de puissances de même exposant\n5 : Mélange"]; 
-};
-
-/**
- * 3A11-2 - Decomposition_facteurs_premiers
- * Décomposer un nombre en facteurs premiers et compter son nombre de diviseurs à partir d'un tableau
- * plusieurs type de nombres à décomposer
- * type 1 : 3 à 5 facteurs premiers max, multiplicités 0,1,2 ou 3 max à préciser
- * type 2 : un produit de deux premiers entre 30 et 100, multiplicité 1 ... suffisamment de possibilités?
- * type 3 : un gros premiers au delà de 1000 et inférieur à 2 000  
- * @author Sébastien Lozano
- */
-
-function Decomposition_facteurs_premiers() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.sup = 1;
-	this.titre = "Décomposition en facteurs premiers d'un entier";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = `À l'aide de la calculatrice, décomposer pas à pas les nombres entiers en produit de facteurs premiers.`;
-	//this.consigne += `<br>`;
-	sortie_html ? this.spacing = 3 : this.spacing = 2;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	this.nb_questions = 3;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-	this.liste_packages = `bclogo`;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//this.bouton_aide = '';
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheArithmetique-3A11.pdf", "Aide mémoire sur les nombres premiers (Sébastien Lozano)", "Aide mémoire")
-			this.bouton_aide += modal_video('conteMathsNombresPremiers', 'videos/LesNombresPremiers.mp4', 'Petit conte mathématique - Les Nombres Premiers', 'Intro Vidéo');
-		} else { // sortie LaTeX
-		};
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = [1, 2, 3];
-		type_de_questions_disponibles = shuffle(type_de_questions_disponibles); // on mélange l'ordre des questions
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-
-		let string_rappel = `Cette liste des nombres premiers inférieurs à 100 pourra être utile : <br>` + crible_eratosthene_n(100)[0];
-		for (let k = 1; k < crible_eratosthene_n(100).length; k++) {
-			string_rappel += `, ` + crible_eratosthene_n(100)[k];
-		};
-		string_rappel += `.`;
-
-		this.introduction = warn_message(string_rappel, `nombres`, `Coup de pouce`);
-
-		for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-			switch (type_de_questions) {
-				case 1: // 3 à 5 facteurs premiers max compris entre 0 et 30, de multiplicité 1,2 ou 3 max
-					// on fixe le nombre de facteurs premier entre 3 et 5
-					let nb_de_premiers = randint(3, 5);
-					// on fixe la limite pour le choix des premiers
-					let max_premier = 11;
-					// on fixe le rang max pour le choix des premiers
-					let rg_max = crible_eratosthene_n(max_premier).length - 1;
-					// on choisit les rangs pour les nombres premiers
-					let tab_rangs = [];
-					let tab_rangs_exclus = [];
-					for (let k = 0; k < (nb_de_premiers); k++) {
-						for (let m = 0; m < k; m++) {
-							tab_rangs_exclus.push(tab_rangs[m]);
-						}
-						tab_rangs[k] = randint(0, rg_max, tab_rangs_exclus);
-					};
-					// on choisit les premiers
-					let tab_premiers = [];
-					for (let k = 0; k < tab_rangs.length; k++) {
-						tab_premiers[k] = crible_eratosthene_n(max_premier)[tab_rangs[k]];
-					};
-					// on range les facteurs premiers dans l'ordre croissant
-					tab_premiers.sort(function (a, b) {
-						return a - b;
-					});
-					// on choisit les multiplicités
-					let tab_multiplicites = [];
-					for (let k = 0; k < tab_rangs.length; k++) {
-						tab_multiplicites[k] = randint(1, 2);
-					};
-					// yapluka écrire le nombre dans l'énoncé et sa décomposition dans la correction
-					texte = `À l'aide de la calculatrice, décomposer `;
-					let nombre_a_decomposer = 1;
-					for (let k = 0; k < tab_rangs.length; k++) {
-						for (let m = 0; m < tab_multiplicites[k]; m++) {
-							nombre_a_decomposer = nombre_a_decomposer * tab_premiers[k];
-						};
-					};
-					let racine_premier_1 = Math.trunc(Math.sqrt(nombre_a_decomposer));
-					texte += `$${tex_nombre(nombre_a_decomposer)}$ en produit de facteurs premiers.`;
-					// correction						
-					texte_corr = `Nous allons successivement tester la divisibilité de $${tex_nombre(nombre_a_decomposer)}$ par tous les nombres premiers inférieurs à `;
-					texte_corr += `$${tex_nombre(nombre_a_decomposer)}$ en commençant par 2, 3, 5, 7, ...<br>`;
-					texte_corr = `Il est suffisant de tester la divisibilité de $${tex_nombre(nombre_a_decomposer)}$ par tous les nombres premiers inférieurs ou égaux à $\\sqrt{${tex_nombre(nombre_a_decomposer)}}$ c'est à dire inférieurs à $${tex_nombre(racine_premier_1)}$.<br>`;
-					texte_corr += `Ce sont les nombres de la liste : <br>`;
-					texte_corr += crible_eratosthene_n(racine_premier_1)[0] + ` ; `;
-					for (let k = 1; k < crible_eratosthene_n(racine_premier_1).length; k++) {
-						texte_corr += crible_eratosthene_n(racine_premier_1)[k];
-						if (k != crible_eratosthene_n(racine_premier_1).length - 1) {
-							texte_corr += ` ; `;
-						} else {
-							texte_corr += `.`;
-						}
-						if (k % 15 == 0) {
-							texte_corr += `<br>`;
-						}
-					};
-					texte_corr += `<br>`;
-					var liste_facteurs_premiers = obtenir_liste_facteurs_premiers(nombre_a_decomposer);
-					var quotient_intermediaire = nombre_a_decomposer;
-					for (let k = 0; k < liste_facteurs_premiers.length; k++) {
-						texte_corr += `$${tex_nombre(quotient_intermediaire)}\\div${mise_en_evidence(liste_facteurs_premiers[k])} = ${tex_nombre(quotient_intermediaire / liste_facteurs_premiers[k])}$<br>`;
-						quotient_intermediaire = quotient_intermediaire / liste_facteurs_premiers[k];
-					};
-					texte_corr += `Finalement on obtient la décomposition suivante : $ ${tex_nombre(nombre_a_decomposer)} = `;
-					if (tab_multiplicites[0] == 1) {
-						texte_corr += `${tab_premiers[0]}`;
-					} else {
-						texte_corr += `${tab_premiers[0]}^{${tab_multiplicites[0]}}`;
-					};
-					for (let k = 1; k < tab_premiers.length; k++) {
-						if (tab_multiplicites[k] == 1) {
-							texte_corr += `\\times ${tab_premiers[k]}`;
-						} else {
-							texte_corr += `\\times ${tab_premiers[k]}^{${tab_multiplicites[k]}}`;
-						};
-					};
-					texte_corr += `$`;
-					break;
-				case 2: // deux premiers compris entre 30 et 100 de multiplicité 1
-					// on choisit un rang différent pour chaque premier entre 30 et 100
-					let r1 = randint(0, premiers_entre_bornes(30, 100).length - 1);
-					let r2 = randint(0, premiers_entre_bornes(30, 100).length - 1, r1);
-					let premier1 = premiers_entre_bornes(30, 100)[r1];
-					let premier2 = premiers_entre_bornes(30, 100)[r2];
-					if (premier1 > premier2) { // on inverse p1 et p2 si p1 est supérieur à p2
-						let p = premier1;
-						premier1 = premier2;
-						premier2 = p;
-					};
-					texte = `À l'aide de la calculatrice, décomposer $${tex_nombre(premier1 * premier2)}$ en produit de facteurs premiers.`;
-					let racine_prem = Math.trunc(Math.sqrt(premier1 * premier2));
-					texte_corr = `Il est suffisant de tester la divisibilité de $${tex_nombre(premier1 * premier2)}$ par tous les nombres premiers inférieurs ou égaux à $\\sqrt{${tex_nombre(premier1 * premier2)}}$ c'est à dire inférieurs à $${tex_nombre(racine_prem)}$.<br>`;
-					texte_corr += `Ce sont les nombres de la liste suivante : <br>$`;
-					texte_corr += crible_eratosthene_n(racine_prem)[0];
-					for (let k = 1; k < crible_eratosthene_n(racine_prem).length; k++) {
-						texte_corr += `; ` + crible_eratosthene_n(racine_prem)[k];
-					};
-					texte_corr += `.$<br>`;
-					liste_facteurs_premiers = obtenir_liste_facteurs_premiers(premier1 * premier2);
-					quotient_intermediaire = premier1 * premier2;
-					for (let k = 0; k < liste_facteurs_premiers.length; k++) {
-						texte_corr += `$${tex_nombre(quotient_intermediaire)}\\div${mise_en_evidence(liste_facteurs_premiers[k])} = ${tex_nombre(quotient_intermediaire / liste_facteurs_premiers[k])}$<br>`;
-						quotient_intermediaire = quotient_intermediaire / liste_facteurs_premiers[k];
-					};
-					texte_corr += ` D'où $${tex_nombre(premier1 * premier2)} = ${tex_nombre(premier1)}\\times${tex_nombre(premier2)}$.`;
-					break;
-				case 3: // un gros premier entre 1000 et 2000			
-					// on choisit un rang pour le nombre premier entre 1000 et 2000
-					let r = randint(0, premiers_entre_bornes(1000, 2000).length - 1);
-					let premier = premiers_entre_bornes(1000, 2000)[r];
-					let racine_premier = Math.trunc(Math.sqrt(premier));
-					texte = `À l'aide de la calculatrice, décomposer $${tex_nombre(premier)}$ en produit de facteurs premiers.`;
-					texte_corr = `En testant la divisibilité de $${tex_nombre(premier)}$ par tous les nombres premiers inférieurs ou égaux à $${racine_premier}$`;
-					texte_corr += ` c'est à dire les nombre de la liste $`;
-					texte_corr += crible_eratosthene_n(racine_premier)[0];
-					for (let k = 1; k < crible_eratosthene_n(racine_premier).length; k++) {
-						texte_corr += `; ` + crible_eratosthene_n(racine_premier)[k];
-					};
-					texte_corr += `$, `;
-					texte_corr += `on se rend compte que $${tex_nombre(premier)}$ est un nombre premier donc `;
-					texte_corr += `$${tex_nombre(premier)} = ${tex_nombre(premier)}$.`;
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	//this.besoin_formulaire_numerique = ['Règle à travailler',5,"1 : Produit de deux puissances de même base\n2 : Quotient de deux puissances de même base\n3 : Puissance de puissance\n4 : Produit de puissances de même exposant\n5 : Mélange"]; 
-};
-
-/**
- * 3A11-3 - Lister/Compter les diviseurs d'un entier à partir de sa decomposition en facteurs premiers 
- * @author Sébastien Lozano
- */
-
-function Lister_Diviseurs_Par_Decomposition_facteurs_premiers() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.sup = 1;
-	this.titre = "Compter/lister les diviseurs d'un entier à partir de sa décomposition en facteurs premiers.";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = `Sans la calculatrice, compter/lister les diviseurs d'un entier à partir de sa décomposition en facteurs premiers.`;
-	//this.consigne += `<br>`;
-	sortie_html ? this.spacing = 2 : this.spacing = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	this.nb_questions = 2;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//this.bouton_aide = '';
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheArithmetique-3A11.pdf", "Aide mémoire sur les nombres premiers (Sébastien Lozano)", "Aide mémoire")
-			this.bouton_aide += modal_video('conteMathsNombresPremiers', 'videos/LesNombresPremiers.mp4', 'Petit conte mathématique - Les Nombres Premiers', 'Intro Vidéo');
-		} else { // sortie LaTeX
-		};
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = [1];
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-
-		for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-
-			switch (type_de_questions) {
-				case 1: // lister/compter les diviseurs d'un entier à partir de sa décomposition en facteurs premiers			
-					texte = `Lister/compter les diviseurs d'un entier à partir de sa décomposition en facteurs premiers`;
-					//let premiers_dispos = premiers_entre_bornes(2,11);
-					// on fixe le nombre de facteurs premier à 3
-					let nb_de_premiers_b = randint(3, 3);
-					// on fixe la limite pour le choix des premiers
-					let max_premier_b = 11;
-					// on fixe le rang max pour le choix des premiers
-					let rg_max_b = crible_eratosthene_n(max_premier_b).length - 1;
-					// on choisit les rangs pour les nombres premiers
-					let tab_rangs_b = [];
-					let tab_rangs_exclus_b = [];
-					for (let k = 0; k < (nb_de_premiers_b); k++) {
-						for (let m = 0; m < k; m++) {
-							tab_rangs_exclus_b.push(tab_rangs_b[m]);
-						}
-						tab_rangs_b[k] = randint(0, rg_max_b, tab_rangs_exclus_b);
-					};
-					// on choisit les premiers
-					let tab_premiers_b = [];
-					for (let k = 0; k < tab_rangs_b.length; k++) {
-						tab_premiers_b[k] = crible_eratosthene_n(max_premier_b)[tab_rangs_b[k]];
-					};
-					// on range les facteurs premiers dans l'ordre croissant
-					tab_premiers_b.sort(function (a, b) {
-						return a - b;
-					});
-					// on choisit les multiplicités
-					let tab_multiplicites_b = [];
-					for (let k = 0; k < tab_rangs_b.length; k++) {
-						tab_multiplicites_b[k] = randint(1, 2);
-					};
-					texte = ``;
-					let nombre_a_decomposer_b = 1;
-					for (let k = 0; k < tab_rangs_b.length; k++) {
-						for (let m = 0; m < tab_multiplicites_b[k]; m++) {
-							nombre_a_decomposer_b = nombre_a_decomposer_b * tab_premiers_b[k];
-						};
-					};
-					texte += `La décomposition en facteurs premiers de $${tex_nombre(nombre_a_decomposer_b)}$ est : $`;
-					if (tab_multiplicites_b[0] == 1) {
-						texte += `${tab_premiers_b[0]}`;
-					} else {
-						texte += `${tab_premiers_b[0]}^{${tab_multiplicites_b[0]}}`;
-					};
-					for (let k = 1; k < tab_premiers_b.length; k++) {
-						if (tab_multiplicites_b[k] == 1) {
-							texte += `\\times ${tab_premiers_b[k]}`;
-						} else {
-							texte += `\\times ${tab_premiers_b[k]}^{${tab_multiplicites_b[k]}}`;
-						};
-					};
-					texte += `$, <br>`;
-					texte += num_alpha(0) + ` Compléter le tableau ci-dessous.`;
-					if (!sortie_html) {
-						texte += `$\\medskip$`;
-					};
-					// on crée le tableau des entetes de lignes et des colonnes
-					let ent_lignes = [];
-					let contenu_lignes = [];
-					let ent_colonnes = [`\\times`];
-					// les entetes des lignes
-					for (let k = 0; k < tab_multiplicites_b[0] + 1; k++) {
-						ent_lignes.push(`\\phantom{plusLarge}` + tab_premiers_b[0] + `^{` + k + `}\\phantom{plusLarge}`);
-					};
-					// les entetes des colonnes 
-					for (let m = 0; m < tab_multiplicites_b[1] + 1; m++) {
-						for (let l = 0; l < tab_multiplicites_b[2] + 1; l++) {
-							ent_colonnes.push(tab_premiers_b[1] + `^{` + m + `}\\times` + tab_premiers_b[2] + `^{` + l + `}`);
-						};
-					};
-					// tableau pour la permutation circulaire
-					let tab_temp;
-					// on y affecte les lignes
-					tab_temp = ent_lignes;
-					// on supprime le x de l'entete des colonnes
-					ent_colonnes.shift();
-					// on affecte ça aux lignes;
-					ent_lignes = ent_colonnes;
-					// on remet le x en colonnes et on ajoute le reste
-					ent_colonnes = [`\\times`].concat(tab_temp);
-					// le contenu des lignes
-					for (let l = 0; l < (tab_multiplicites_b[0] + 1); l++) {
-						for (let c = 1; c < (tab_multiplicites_b[1] + 1) * (tab_multiplicites_b[2] + 1) + 1; c++) {
-							//contenu_lignes.push(`l : `+l+`, c : `+Number(c));
-							contenu_lignes.push(``);
-						};
-					};
-					texte += `<br>`;
-					texte += tab_C_L(ent_colonnes, ent_lignes, contenu_lignes);
-					if (!sortie_html) {
-						texte += `$\\medskip$`;
-					};
-					texte += `<br>`;
-					texte += num_alpha(1) + ` En déduire le nombre de diviseurs de $${tex_nombre(nombre_a_decomposer_b)}$.<br>`;
-					texte += num_alpha(2) + ` Enfin, dresser la liste des diviseurs de $${tex_nombre(nombre_a_decomposer_b)}$.<br>`;
-
-					// correction
-					texte_corr = `Avec la décomposition en facteurs premiers de $${tex_nombre(nombre_a_decomposer_b)}$ qui est : $`;
-					if (tab_multiplicites_b[0] == 1) {
-						texte_corr += `${tab_premiers_b[0]}`;
-					} else {
-						texte_corr += `${tab_premiers_b[0]}^{${tab_multiplicites_b[0]}}`;
-					};
-					for (let k = 1; k < tab_premiers_b.length; k++) {
-						if (tab_multiplicites_b[k] == 1) {
-							texte_corr += `\\times ${tab_premiers_b[k]}`;
-						} else {
-							texte_corr += `\\times ${tab_premiers_b[k]}^{${tab_multiplicites_b[k]}}`;
-						};
-					};
-					texte_corr += `$, <br>`;
-					texte_corr += num_alpha(0) + ` Le tableau donne :`;
-					// on crée le tableau des entetes de lignes et des colonnes
-					let ent_lignes_corr = [];
-					let ent_lignes_corr_res = [];
-					let contenu_lignes_corr = [];
-					//let contenu_lignes_corr_res = [];
-					let ent_colonnes_corr = [`\\times`];
-					let ent_colonnes_corr_res = [1];
-					// les entetes des lignes
-					for (let k = 0; k < tab_multiplicites_b[0] + 1; k++) {
-						ent_lignes_corr.push(tab_premiers_b[0] + `^{` + k + `}`);
-						ent_lignes_corr_res.push(tab_premiers_b[0] ** k);
-					};
-					// les entetes des colonnes 
-					for (let m = 0; m < tab_multiplicites_b[1] + 1; m++) {
-						for (let l = 0; l < tab_multiplicites_b[2] + 1; l++) {
-							ent_colonnes_corr.push(tab_premiers_b[1] + `^{` + m + `}\\times` + tab_premiers_b[2] + `^{` + l + `}`);
-							ent_colonnes_corr_res.push(tab_premiers_b[1] ** m * tab_premiers_b[2] ** l);
-						};
-					};
-					// tableaux pour les permutations circulaires
-					let tab_temp_corr;
-					let tab1_temp_corr;
-					// on y affecte les lignes
-					tab_temp_corr = ent_lignes_corr;
-					tab1_temp_corr = ent_lignes_corr_res;
-					// on supprime le x de l'entete des colonnes
-					ent_colonnes_corr.shift();
-					ent_colonnes_corr_res.shift();
-					// on affecte ça aux lignes;
-					ent_lignes_corr = ent_colonnes_corr;
-					ent_lignes_corr_res = ent_colonnes_corr_res;
-					// on remet le x en colonnes et on ajoute le reste
-					ent_colonnes_corr = [`\\times`].concat(tab_temp_corr);
-					ent_colonnes_corr_res = [1].concat(tab1_temp_corr);
-					// le contenu des lignes
-					for (let l = 0; l < (tab_multiplicites_b[1] + 1) * (tab_multiplicites_b[2] + 1) + 1; l++) {
-						for (let c = 1; c < (tab_multiplicites_b[0] + 2); c++) {
-							//contenu_lignes_corr.push(`l : `+l+`, c : `+Number(c));								
-							contenu_lignes_corr.push(ent_lignes_corr[l] + `\\times` + ent_colonnes_corr[c] + `=` + mise_en_evidence(tex_nombre(ent_lignes_corr_res[l] * ent_colonnes_corr_res[c])));
-						};
-					};
-					texte_corr += `<br>`;
-					texte_corr += tab_C_L(ent_colonnes_corr, ent_lignes_corr, contenu_lignes_corr);
-					texte_corr += `<br>`;
-					texte_corr += num_alpha(1) + ` $${tex_nombre(nombre_a_decomposer_b)}$ a donc `;
-					texte_corr += `$(${tab_multiplicites_b[0]}+1)\\times(${tab_multiplicites_b[1]}+1)\\times(${tab_multiplicites_b[2]}+1) = `;
-					texte_corr += `${tab_multiplicites_b[0] + 1}\\times${tab_multiplicites_b[1] + 1}\\times${tab_multiplicites_b[2] + 1} = `;
-					texte_corr += `${(tab_multiplicites_b[0] + 1) * (tab_multiplicites_b[1] + 1) * (tab_multiplicites_b[2] + 1)}$ diviseurs.<br>`;
-					texte_corr += `En effet, dans la décomposition apparait : `;
-					texte_corr += ` <br> - Le facteur premier $${tab_premiers_b[0]}$ avec la multiplicité $${tab_multiplicites_b[0]}$`;
-					texte_corr += `, le facteur $${tab_premiers_b[0]}$ apparait donc sous les formes : `;
-					for (let k = 0; k < tab_multiplicites_b[0]; k++) {
-						texte_corr += `$${tab_premiers_b[0]}^{` + k + `}$ ou `;
-					};
-					texte_corr += `$${tab_premiers_b[0]}^{` + tab_multiplicites_b[0] + `}$ d'où le facteur $(${tab_multiplicites_b[0]}+1)$.`;
-
-					texte_corr += ` <br> - Le facteur premier $${tab_premiers_b[1]}$ avec la multiplicité $${tab_multiplicites_b[1]}$`;
-					texte_corr += `, le facteur $${tab_premiers_b[1]}$ apparait donc sous les formes : `;
-					for (let k = 0; k < tab_multiplicites_b[1]; k++) {
-						texte_corr += `$${tab_premiers_b[1]}^{` + k + `}$ ou `;
-					};
-					texte_corr += `$${tab_premiers_b[1]}^{` + tab_multiplicites_b[1] + `}$ d'où le facteur $(${tab_multiplicites_b[1]}+1)$.`;
-
-					texte_corr += ` <br> - Le facteur premier $${tab_premiers_b[2]}$ avec la multiplicité $${tab_multiplicites_b[2]}$`;
-					texte_corr += `, le facteur $${tab_premiers_b[2]}$ apparait donc sous les formes : `;
-					for (let k = 0; k < tab_multiplicites_b[2]; k++) {
-						texte_corr += `$${tab_premiers_b[2]}^{` + k + `}$ ou `;
-					};
-					texte_corr += `$${tab_premiers_b[2]}^{` + tab_multiplicites_b[2] + `}$ d'où le facteur $(${tab_multiplicites_b[2]}+1)$.`;
-					texte_corr += `<br>`;
-					texte_corr += num_alpha(2) + ` Enfin, voici la liste des $${(tab_multiplicites_b[0] + 1) * (tab_multiplicites_b[1] + 1) * (tab_multiplicites_b[2] + 1)}$ diviseurs de $${tex_nombre(nombre_a_decomposer_b)}$ issus du tableau ci-dessus : `;
-					texte_corr += `$1`;
-					for (let w = 1; w < liste_diviseurs(nombre_a_decomposer_b).length; w++) {
-						texte_corr += `\\text{ ; }` + tex_nombre(liste_diviseurs(nombre_a_decomposer_b)[w]);
-					};
-					texte_corr += `.$`;
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	//this.besoin_formulaire_numerique = ['Règle à travailler',5,"1 : Produit de deux puissances de même base\n2 : Quotient de deux puissances de même base\n3 : Puissance de puissance\n4 : Produit de puissances de même exposant\n5 : Mélange"]; 
-};
-
-/**
- * 3A12 - Fractions irreductibles
- * @author Sébastien Lozano
- */
-
-function Fractions_irreductibles() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.sup = 1;
-	this.titre = "Fractions irréductibles";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = `Rendre irréductible une fraction et son inverse à partir des décompositons en produit de facteurs premiers.`;
-	//this.consigne += `<br>`;
-	sortie_html ? this.spacing = 4 : this.spacing = 2;
-	sortie_html ? this.spacing_corr = 4 : this.spacing_corr = 2;
-	this.nb_questions = 1;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-	this.liste_packages = `bclogo`;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//this.bouton_aide = '';
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheArithmetique-3A12.pdf", "Aide mémoire sur les fonctions (Sébastien Lozano)", "Aide mémoire")
-			//this.bouton_aide += modal_video('conteMathsNombresPremiers','videos/LesNombresPremiers.mp4','Petit conte mathématique','Intro Vidéo');
-		} else { // sortie LaTeX
-		};
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-
-		//let type_de_questions_disponibles = [1,2,3,4];
-		let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-
-		this.introduction = warn_message(`À la question ` + num_alpha(3) + ` une observation judicieuse et argumentée pourra faire gagner du temps!`, `nombres`, `Coup de pouce`);
-
-		for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-			var nb_div_prem_communs; // nombre de diviseurs premiers communs
-			var candidats_premiers_communs; // tableau des candidats premiers communs
-			var premiers_communs; // tableau des diviseurs premiers communs
-			var multiplicites_premiers_communs; // tableau des multiplicités des diviseurs premiers communs 
-			var r; // tableau pour le choix des rangs des diviseurs premiers communs
-			var r_exclus; // tableau pour la boucle de creation de r				
-			var nb1_dist; // diviseur unique du premier nombre 
-			var nb2_dist; // divisuer unique du second nombre
-			var r_ex; // pour exlcure le rang de nb1
-			var tab_nb1; // tableau pour les diviseurs de nb1
-			var tab_nb2; // tableau pour les diviseurs de nb2
-			var multiplicites_nb1;
-			var multiplicites_nb2;
-			var nb1; // nbre 1
-			var nb2; // nbre 2
-
-			// on fixe le tableau de choix
-			candidats_premiers_communs = premiers_entre_bornes(2, 13);
-			// on fixe le nombre de divisuers premiers communs
-			nb_div_prem_communs = 4;
-			// on initialise le tableau des diviseurs premiers communs
-			premiers_communs = [];
-			// on initialise le tableau des rangs
-			r = [];
-			// on initialise le tableau des rangs déjà choisis
-			r_exclus = [];
-			// on complète le tableau des rangs des rangs des diviseurs premiers choisis
-			for (let k = 0; k < nb_div_prem_communs; k++) {
-				for (let m = 0; m < k; m++) {
-					r_exclus.push(r[m]);
-				};
-				r[k] = randint(0, candidats_premiers_communs.length - 1, r_exclus);
-			};
-			// on complète le tableau des diviseurs premiers communs
-			for (let k = 0; k < nb_div_prem_communs; k++) {
-				premiers_communs.push(candidats_premiers_communs[r[k]]);
-			};
-			// on initialise et on complète le tableau des multiplicités des diviseurs premiers communs
-			multiplicites_premiers_communs = [];
-			for (let k = 0; k < nb_div_prem_communs; k++) {
-				multiplicites_premiers_communs.push(randint(0, 2));
-			};
-			// on supprime les diviseurs premiers de multiplicité 0 et leur multiplicité
-			let idx = multiplicites_premiers_communs.indexOf(0);
-			while (idx != -1) {
-				premiers_communs.splice(idx, 1);
-				multiplicites_premiers_communs.splice(idx, 1);
-				idx = multiplicites_premiers_communs.indexOf(0);
-			};
-			// on initialise le tableau des diviseurs du premier et du second nombre avec les diviseurs premiers communs
-			tab_nb1 = [];
-			tab_nb2 = [];
-			for (let k = 0; k < premiers_communs.length; k++) {
-				tab_nb1[k] = premiers_communs[k];
-				tab_nb2[k] = premiers_communs[k];
-			};
-			// on initialise les tableaux de multiplicité, ils sont les mêmes mais on pourrait vouloir qu'ils soient différents
-			multiplicites_nb1 = [];
-			multiplicites_nb2 = [];
-			for (let k = 0; k < premiers_communs.length; k++) {
-				multiplicites_nb1[k] = multiplicites_premiers_communs[k];
-				multiplicites_nb2[k] = multiplicites_premiers_communs[k];
-			};
-			// on ajoute un facteur premier distinct pour chaque nombre plus petit que 30
-			r_ex = randint(0, premiers_entre_bornes(2, 30).length - 1);
-			nb1_dist = premiers_entre_bornes(2, 30)[r_ex];
-			nb2_dist = premiers_entre_bornes(2, 30)[randint(0, premiers_entre_bornes(2, 30).length - 1, r_ex)];
-			// on ajoute nb1_dist, nb2_dist dans les tableaux des diviseurs premiers du premier et du second nombre 
-			// nb1
-			let bool = false;
-			let n = 0;
-			while (n < tab_nb1.length && bool != true) {
-				if (nb1_dist == tab_nb1[n]) {// si le diviseur premier est déjà présent on incrémente sa multiplicité
-					multiplicites_nb1[n]++;
-					bool = true;
-				};
-				n++;
-			};
-			// on teste la valeur de sortie de bool et on ajoute la nouvelle valeur si necessaire
-			if (!bool) {// il n'est pas présent on l'ajoute avec la multipplicité 1
-				tab_nb1.push(nb1_dist);
-				multiplicites_nb1.push(1);
-				bool = true;
-			};
-			// nb2
-			bool = false;
-			n = 0;
-			while (n < tab_nb2.length && !bool) {
-				if (nb2_dist == tab_nb2[n]) {// si le diviseur premier est déjà présent on incrémente sa multiplicité
-					multiplicites_nb2[n]++;
-					bool = true;
-				};
-				n++;
-			};
-			// on teste la valeur de sortie de bool et on ajoute la nouvelle valeur si necessaire
-			if (!bool) {// il n'est pas présent on l'ajoute avec la multipplicité 1
-				tab_nb2.push(nb2_dist);
-				multiplicites_nb2.push(1);
-				bool = true;
-			};
-			// on crée un tableau associatif à partir des deux tableaux tab_ni et multiplicites_ni
-			let tab_prem_mult_nb1 = [];
-			for (let k = 0; k < tab_nb1.length; k++) {
-				tab_prem_mult_nb1[k] = { 'prem': tab_nb1[k], 'mult': multiplicites_nb1[k] };
-			};
-			let tab_prem_mult_nb2 = [];
-			for (let k = 0; k < tab_nb2.length; k++) {
-				tab_prem_mult_nb2[k] = { 'prem': tab_nb2[k], 'mult': multiplicites_nb2[k] };
-			};
-			// on range selon prem croissant
-			tab_prem_mult_nb1.sort(function (a, b) {
-				return a.prem > b.prem;
-			});
-			tab_prem_mult_nb2.sort(function (a, b) {
-				return a.prem > b.prem;
-			});
-			// on initialise nb1 et nb2 et on les calcule à partir des tableaux 				
-			nb1 = 1;
-			for (let k = 0; k < tab_nb1.length; k++) {
-				nb1 = nb1 * tab_prem_mult_nb1[k].prem ** tab_prem_mult_nb1[k].mult;
-			};
-			nb2 = 1;
-			for (let k = 0; k < tab_nb2.length; k++) {
-				nb2 = nb2 * tab_prem_mult_nb2[k].prem ** tab_prem_mult_nb2[k].mult;
-			};
-
-			switch (type_de_questions) {
-				case 1: // décomposition de A
-					texte = num_alpha(0) + ` Décomposer $A = ${tex_nombre(nb1)}$ en produit de facteurs premiers : `;
-					texte_corr = num_alpha(0) + ` La décomposition en produit de facteurs premier de $A = `;
-					switch (tab_prem_mult_nb1[0].mult) {
-						case 1:
-							texte_corr += `${tab_prem_mult_nb1[0].prem}`;
-							break;
-						default:
-							texte_corr += `${tab_prem_mult_nb1[0].prem}^{${tab_prem_mult_nb1[0].mult}}`;
-							break;
-					};
-					for (let k = 1; k < tab_nb1.length; k++) {
-						switch (tab_prem_mult_nb1[k].mult) {
-							case 1:
-								texte_corr += `\\times${tab_prem_mult_nb1[k].prem}`;
-								break;
-							default:
-								texte_corr += `\\times${tab_prem_mult_nb1[k].prem}^{${tab_prem_mult_nb1[k].mult}}`;
-								break;
-						};
-					};
-					texte_corr += `$.`;
-					//	break;		
-					//case 2 : // décomposition de B 	
-					texte += `<br>` + num_alpha(1) + ` Décomposer $B = ${tex_nombre(nb2)}$ en produit de facteurs premiers : `;
-					texte_corr += `<br>` + num_alpha(1) + ` La décomposition en produit de facteurs premier de $B = `;
-					switch (tab_prem_mult_nb2[0].mult) {
-						case 1:
-							texte_corr += `${tab_prem_mult_nb2[0].prem}`;
-							break;
-						default:
-							texte_corr += `${tab_prem_mult_nb2[0].prem}^{${tab_prem_mult_nb2[0].mult}}`;
-							break;
-					};
-					for (let k = 1; k < tab_nb2.length; k++) {
-						switch (tab_prem_mult_nb2[k].mult) {
-							case 1:
-								texte_corr += `\\times${tab_prem_mult_nb2[k].prem}`;
-								break;
-							default:
-								texte_corr += `\\times${tab_prem_mult_nb2[k].prem}^{${tab_prem_mult_nb2[k].mult}}`;
-								break;
-						};
-					};
-					texte_corr += `$.`;
-					//	break;	
-					//case 3 : // reduction de A sur B 			
-					texte += `<br>` + num_alpha(2) + ` Rendre la fraction $\\dfrac{A}{B} = \\dfrac{${tex_nombre(nb1)}}{${tex_nombre(nb2)}}$ irréductible `;
-					if (sortie_html) {
-						texte += ` à l'aide des décompositions obtenues au ` + num_alpha(0) + ` et au ` + num_alpha(1);
-					} else {
-						texte += ` à l'aide des questions ` + num_alpha(0) + ` et ` + num_alpha(1);
-					};
-					texte_corr += `<br>` + num_alpha(2) + ` $\\dfrac{A}{B} = \\dfrac{${tex_nombre(nb1)}}{${tex_nombre(nb2)}} = `;
-					texte_corr += `\\dfrac{`;
-					texte_corr += `\\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[0] + `}`;
-					for (let k = 1; k < decomp_fact_prem_array(nb1 / nb1_dist).length; k++) {
-						texte_corr += `\\times \\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[k] + `}`
-					};
-					texte_corr += `\\times ${nb1_dist}}{`;
-					texte_corr += `\\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[0] + `}`;
-					for (let k = 1; k < decomp_fact_prem_array(nb1 / nb1_dist).length; k++) {
-						texte_corr += `\\times \\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[k] + `}`
-					};
-					texte_corr += `\\times ${nb2_dist}} = `;
-					texte_corr += `\\dfrac{${nb1_dist}}{${nb2_dist}}$.`;
-					//	break;	
-					//case 4 : // reduction de B sur A 			
-					texte += `<br>` + num_alpha(3) + ` Rendre la fraction $\\dfrac{B}{A} = \\dfrac{${tex_nombre(nb2)}}{${tex_nombre(nb1)}}$ irréductible`;
-					if (sortie_html) {
-						texte += ` à l'aide des décompositions obtenues au ` + num_alpha(0) + ` et au ` + num_alpha(1);
-					} else {
-						texte += ` à l'aide des questions ` + num_alpha(0) + ` et ` + num_alpha(1);
-					};
-					texte_corr += `<br>` + num_alpha(3) + ` $\\dfrac{B}{A}$ est l'inverse de $\\dfrac{A}{B}$ donc $\\dfrac{B}{A} = \\dfrac{${tex_nombre(nb2)}}{${tex_nombre(nb1)}} = `;
-					texte_corr += `\\dfrac{`;
-					texte_corr += `\\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[0] + `}`;
-					for (let k = 1; k < decomp_fact_prem_array(nb1 / nb1_dist).length; k++) {
-						texte_corr += `\\times \\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[k] + `}`
-					};
-					texte_corr += `\\times ${nb2_dist}}{`;
-					texte_corr += `\\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[0] + `}`;
-					for (let k = 1; k < decomp_fact_prem_array(nb1 / nb1_dist).length; k++) {
-						texte_corr += `\\times \\cancel{` + decomp_fact_prem_array(nb1 / nb1_dist)[k] + `}`
-					};
-					texte_corr += `\\times ${nb1_dist}} = `;
-					texte_corr += `\\dfrac{${nb2_dist}}{${nb1_dist}}$.`;
-					//	break;	
-					//case 5 : // calculer le produit A/B x B/A et réduire. Remarque?
-					// texte += `<br>`+num_alpha(4)+` Combien alculer le produit de $\\dfrac{A}{B} = \\dfrac{${tex_nombre(nb1)}}{${tex_nombre(nb2)}}$ et de $\\dfrac{B}{A} = \\dfrac{${tex_nombre(nb2)}}{${tex_nombre(nb1)}}$.`;
-					// texte += `<br>Donner le résultat sous forme de fraction irréductible.`
-					//texte += `<br>`+num_alpha(4)+` Remarque ?`
-					//texte_corr += `<br>`+num_alpha(4)+' corr type 5';
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			};
-			cpt++
-		};
-		liste_de_question_to_contenu(this);
-	};
-	//this.besoin_formulaire_numerique = ['Règle à travailler',5,"1 : Produit de deux puissances de même base\n2 : Quotient de deux puissances de même base\n3 : Puissance de puissance\n4 : Produit de puissances de même exposant\n5 : Mélange"]; 
-};
-
-/**
- * 3A13 - PPCM_Engrenages
- * les deux on besoin de la def partielle serie : stlX
- * pb dans la sortie LaTeX, revoir comment user de la fonction katex_Popup2() pour affichage d'une note hors texte!
- * @author Sébastien Lozano
- */
-
-function PPCM_Engrenages() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.sup = 1;
-	this.titre = "Engrenages";
-	// pas de différence entre la version html et la version latex pour la consigne
-	//this.consigne =`Déterminer au bout de combien de tours les deux roues seront toutes les deux revenues à leur position initiale.`;
-	this.consigne = ``;
-	//this.consigne += `<br>`;
-	sortie_html ? this.spacing = 2 : this.spacing = 2;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	this.nb_questions = 4;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-	this.liste_packages = 'bclogo';
-
-	var num_ex = '3A13'; // pour rendre unique les id des SVG, en cas d'utilisation dans plusieurs exercices y faisant appel
-
-	if (sortie_html) {
-		var pourcentage = '100%'; // pour l'affichage des svg. On a besoin d'une variable globale
-	} else { // sortie LaTeX
-
-	};
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			//this.bouton_aide = '';
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheArithmetique-3A13.pdf", "Aide mémoire sur les fonctions (Sébastien Lozano)", "Aide mémoire")
-			//this.bouton_aide += modal_video('conteMathsNombresPremiers','videos/LesNombresPremiers.mp4','Petit conte mathématique','Intro Vidéo');
-		} else { // sortie LaTeX
-		};
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = [1, 2, 3, 4];
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-		let txt_intro = `Boîte de vitesse, transmission de vélo, de moto, perceuse electrique, tout ça fonctionne avec des engrenages! Mais au fait, comment ça marche, les engrenages?`;
-		if (sortie_html) {
-			txt_intro += warn_message(`Attention, les roues ci-dessous ne comportent pas le nombre de dents de l'énoncé!`, `nombres`, `Coup de pouce`);
-			txt_intro += `<div id="${num_ex}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-			//this.introduction += warn_message(`Attention, les roues ci-dessous ne comportent pas le nombre de dents de l'énoncé!`, `nombres`, `Coup de pouce`);
-			//this.introduction += `<div id="${num_ex}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-			SVG_engrenages(num_ex, 200, 200);
-		};
-
-		this.introduction = lampe_message({
-			titre: `Arithmétique des engrenages`,
-			texte: txt_intro,
-			couleur: `nombres`
-		});
-
-		for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-			if (sortie_html) {
-				let id_unique = `${num_ex}_${i}_${Date.now()}`
-				var id_du_div = `div_svg${numero_de_l_exercice}${id_unique}`;
-				//var id_du_div_corr = `div_svg_corr${numero_de_l_exercice}${id_unique}`;
-			}
-
-			var nb_dents_r1;
-			var nb_dents_r2;
-			let txt_popup = `Étant donnés deux nombres entiers a et b, lorsque le plus petit multiple commun à $a$ et $b$ vaut $a \\times b$ ( $ppcm(a,b)=a\\times b$ ), on dit que `;
-			//txt_popup += texte_gras('les nombres a et b sont premiers entre eux');
-			if (sortie_html) {
-				txt_popup += '<b>les nombres a et b sont premiers entre eux</b>';
-			} else {
-				txt_popup += '$\\textbf{les nombres a et b sont premiers entre eux}$';
-			};
-
-			switch (type_de_questions) {
-				case 1: // avec de petits nombres on calcule les mutliples
-					nb_dents_r1 = randint(5, 30);
-					nb_dents_r2 = randint(5, 30, nb_dents_r1);
-					texte = `La roue n$\\degree$1 possède $${nb_dents_r1}$ dents et la roue n$\\degree$2 a $${nb_dents_r2}$ dents.`;
-					texte += `<br>` + num_alpha(0) + ` Écrire la liste des multiples de $${nb_dents_r1}$ et de $${nb_dents_r2}$.`
-					if (ppcm(nb_dents_r1, nb_dents_r2) == (nb_dents_r1 * nb_dents_r2)) {
-						texte += `<br>Pourquoi peut-on en déduire que ${nb_dents_r1} et ${nb_dents_r2} sont des `;
-						// let txt_popup = `Étant donnés deux nombres entiers a et b, lorsque $ppcm(a,b)=a\\times b$, on dit que `;
-						// //txt_popup += texte_gras('les nombres a et b sont premiers entre eux');
-						// if (sortie_html) {
-						// 	txt_popup += '<b>les nombres a et b sont premiers entre eux</b>';
-						// } else {
-						// 	txt_popup += '$\\textbf{les nombres a et b sont premiers entre eux}$';
-						// };
-						//${texte_gras('les nombres a et b sont premiers entre eux')}.`;
-						texte += katex_Popup2(
-							numero_de_l_exercice + 1,
-							1,
-							"nombres premiers entre eux ?",
-							`Définition : Nombres premiers entre eux`,
-							txt_popup
-						);
-					};
-					texte += `<br>` + num_alpha(1) + ` En déduire le nombre de tours de chaque roue avant le retour à leur position initiale.`
-					texte_corr = num_alpha(0) + ` Liste des premiers multiples de $${nb_dents_r1}$ : <br>`;
-					// on va faire en sorte de toujours avoir un nombre de multiples multiple de 5
-					let nb_marge = 5 - (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1) % 5;
-					let k_max = (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1 + nb_marge);
-					for (let k = 1; k < k_max + 1; k++) {
-						texte_corr += `$${k}\\times${nb_dents_r1} = `;
-						if (k == (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1)) {
-							texte_corr += mise_en_evidence(tex_nombre(k * nb_dents_r1));
-							texte_corr += `$ ; `;
-						} else {
-							texte_corr += `${tex_nombre(k * nb_dents_r1)}$ ; `;
-						};
-						if (k % 5 == 0) {
-							texte_corr += `<br>`;
-						}
-					};
-					texte_corr += `$\\ldots$ `;
-					texte_corr += `<br>`;
-					texte_corr += ` Liste des premiers multiples de ${nb_dents_r2} : <br>`;
-					// on va faire en sorte de toujours avoir un nombre de multiples multiple de 5
-					nb_marge = 5 - (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2) % 5;
-					k_max = (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2 + nb_marge);
-					for (let k = 1; k < k_max + 1; k++) {
-						texte_corr += `$${k}\\times${nb_dents_r2} = `;
-						if (k == (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2)) {
-							texte_corr += mise_en_evidence(tex_nombre(k * nb_dents_r2));
-							texte_corr += `$ ; `;
-						} else {
-							texte_corr += `${tex_nombre(k * nb_dents_r2)}$ ; `;
-						};
-						if (k % 5 == 0) {
-							texte_corr += `<br>`;
-						}
-					};
-					texte_corr += `$\\ldots$ `;
-					texte_corr += `<br>`;
-					texte_corr += `Le plus petit multiple commun à $${nb_dents_r1}$ et $${nb_dents_r2}$ vaut donc $ppcm(${nb_dents_r1},${nb_dents_r2}) = ${ppcm(nb_dents_r1, nb_dents_r2)}$.`;
-					texte_corr += `<br>`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) == (nb_dents_r1 * nb_dents_r2)) {
-						texte_corr += `Le $ppcm(` + nb_dents_r1 + `;` + nb_dents_r2 + `)=` + nb_dents_r1 + `\\times` + nb_dents_r2 + `$ donc $${nb_dents_r1}$ et $${nb_dents_r2}$ sont des `;
-						texte_corr += katex_Popup2(
-							numero_de_l_exercice + 2,
-							1,
-							"nombres premiers entre eux.",
-							`Définition : Nombres premiers entre eux`,
-							txt_popup//`Étant donnés deux nombres entiers a et b, lorsque $ppcm(a,b)=a\\times b$, on dit que ${texte_gras('les nombres a et b sont premiers entre eux')}.`
-						);
-					};
-					texte_corr += `<br><br>` + num_alpha(1) + ` Chaque roue doit tourner de $ppcm(${nb_dents_r1},${nb_dents_r2})=${tex_nombre(ppcm(nb_dents_r1, nb_dents_r2))}$ dents.`;
-					texte_corr += `<br> Cela correspond à $(${ppcm(nb_dents_r1, nb_dents_r2)}\\text{ dents})\\div (${nb_dents_r1}\\text{ dents/tour}) = ${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1}$`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1 == 1) {
-						texte_corr += ` tour `;
-					} else {
-						texte_corr += ` tours `;
-					};
-					texte_corr += `pour la roue n$\\degree$1.`
-					texte_corr += `<br>Cela correspond à $(${ppcm(nb_dents_r1, nb_dents_r2)}\\text{ dents})\\div (${nb_dents_r2}\\text{ dents/tour}) = ${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2}$`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2 == 1) {
-						texte_corr += ` tour `;
-					} else {
-						texte_corr += ` tours `;
-					};
-					texte_corr += `pour la roue n$\\degree$2.`
-					break;
-				case 2: // avec de plus grands nombre, c'est mieux de décomposer en facteurs premiers
-					nb_dents_r1 = randint(31, 80);
-					nb_dents_r2 = randint(31, 80, nb_dents_r1);
-					texte = `La roue n$\\degree$1 possède $${nb_dents_r1}$ dents et la roue n$\\degree$2 a $${nb_dents_r2}$ dents.`;
-					texte += `<br>` + num_alpha(0) + ` Décomposer $${nb_dents_r1}$ et $${nb_dents_r2}$ en produit de facteurs premiers.`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) == (nb_dents_r1 * nb_dents_r2)) {
-						texte += `<br>Pourquoi peut-on en déduire que ${nb_dents_r1} et ${nb_dents_r2} sont des `;
-						texte += katex_Popup2(
-							numero_de_l_exercice + 3,
-							1,
-							"nombres premiers entre eux",
-							`Définition : Nombres premiers entre eux`,
-							txt_popup//`Étant donnés deux nombres entiers a et b, lorsque $ppcm(a,b)=a\\times b$, on dit que ${texte_gras('les nombres a et b sont premiers entre eux')}.`
-						);
-					};
-					texte += `<br>` + num_alpha(1) + ` En déduire le nombre de tours de chaque roue avant le retour à leur position initiale.`;
-					texte_corr = `Pour un nombre de dents plus élevé, il est plus commode d'utiliser les décompositions en produit de facteurs premiers.`
-					texte_corr += `<br>` + num_alpha(0) + ` Décomposition de $${nb_dents_r1}$ en produit de facteurs premiers :  $${nb_dents_r1} = ${decomposition_facteurs_premiers(nb_dents_r1)}$.`;
-					texte_corr += `<br> Décomposition de $${nb_dents_r2}$ en produit de facteurs premiers :  $${nb_dents_r2} = ${decomposition_facteurs_premiers(nb_dents_r2)}$.`;
-					texte_corr += `<br> D'où $ppcm(${nb_dents_r1},${nb_dents_r2})= ${decomposition_facteurs_premiers(ppcm(nb_dents_r1, nb_dents_r2))}$.<br>`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) == (nb_dents_r1 * nb_dents_r2)) {
-						texte_corr += `Le $ppcm(` + nb_dents_r1 + `;` + nb_dents_r2 + `)=` + nb_dents_r1 + `\\times` + nb_dents_r2 + `$ donc $${nb_dents_r1}$ et $${nb_dents_r2}$ sont des `;
-						texte_corr += katex_Popup2(
-							numero_de_l_exercice + 4,
-							1,
-							"nombres premiers entre eux.",
-							`Définition : Nombres premiers entre eux`,
-							txt_popup//`Étant donnés deux nombres entiers a et b, lorsque $ppcm(a,b)=a\\times b$, on dit que ${texte_gras('les nombres a et b sont premiers entre eux')}.`
-						);
-					};
-					texte_corr += `<br><br>` + num_alpha(1) + ` Chaque roue doit tourner de $ppcm(${nb_dents_r1},${nb_dents_r2})=${tex_nombre(ppcm(nb_dents_r1, nb_dents_r2))}$ dents.`;
-					texte_corr += `<br> Cela correspond à $(${tex_nombre(ppcm(nb_dents_r1, nb_dents_r2))}\\text{ dents})\\div (${nb_dents_r1}\\text{ dents/tour}) = ${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1}$`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1 == 1) {
-						texte_corr += ` tour `;
-					} else {
-						texte_corr += ` tours `;
-					};
-					texte_corr += `pour la roue n$\\degree$1.`;
-					texte_corr += `<br> Cela correspond à $(${tex_nombre(ppcm(nb_dents_r1, nb_dents_r2))}\\text{ dents})\\div (${nb_dents_r2}\\text{ dents/tour}) = ${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2}$`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2 == 1) {
-						texte_corr += ` tour `;
-					} else {
-						texte_corr += ` tours `;
-					};
-					texte_corr += `pour la roue n$\\degree$2.`
-					break;
-				case 3: // déterminer le nombre de dents d'une roue connaissant l'autre et le nombre de tours necessaires à la re-synchro
-					nb_dents_r1 = randint(5, 80);
-					nb_dents_r2 = randint(5, 80, nb_dents_r1);
-					texte = `La roue n$\\degree$2 a maintenant $${nb_dents_r2}$ dents.`;
-					texte += ` Déterminer le nombre de dents de la roue n$\\degree$1 qui ferait $${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1}$ `;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1 == 1) {
-						texte += ` tour `;
-					} else {
-						texte += ` tours `;
-					};
-					texte += ` pendant que la roue n$\\degree$2 en fait $${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2}$.`
-					texte_corr = `Puisque la roue n$\\degree$2, qui a $${nb_dents_r2}$ dents, fait $${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2}$ `;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r2 == 1) {
-						texte_corr += ` tour `;
-					} else {
-						texte_corr += ` tours `;
-					};
-					texte_corr += `, cela représente $${tex_nombre(ppcm(nb_dents_r1, nb_dents_r2))}$ dents.`;
-					texte_corr += `<br>La roue n$\\degree$1 doit donc aussi tourner de $${tex_nombre(ppcm(nb_dents_r1, nb_dents_r2))}$ dents, ceci en $${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1}$ `;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1 == 1) {
-						texte_corr += ` tour `;
-					} else {
-						texte_corr += ` tours `;
-					};
-					texte_corr += `.`;
-					texte_corr += `<br> on obtient donc $(${tex_nombre(ppcm(nb_dents_r1, nb_dents_r2))}\\text{ dents})\\div (${ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1}\\text{`;
-					if (ppcm(nb_dents_r1, nb_dents_r2) / nb_dents_r1 == 1) {
-						texte_corr += ` tour `;
-					} else {
-						texte_corr += ` tours `;
-					};
-					texte_corr += `}) = ${nb_dents_r1} \\text{ dents/tour}.$`
-					texte_corr += `<br>La roue n$\\degree$1 a donc : $${nb_dents_r1}$ dents.`;
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	//this.besoin_formulaire_numerique = ['Règle à travailler',5,"1 : Produit de deux puissances de même base\n2 : Quotient de deux puissances de même base\n3 : Puissance de puissance\n4 : Produit de puissances de même exposant\n5 : Mélange"]; 
-};
+;
 
 
 /**
