+++ modules/outils.js
@@ -2597,9 +2597,9 @@ export function  SVG_reperage_sur_un_axe(id_du_div,origine,longueur,pas1,pas2,po
 			let droite = mon_svg.line(100, 50, 750, 50)	
 			droite.stroke({ color: 'black', width: 2, linecap: 'round' })
 			// Graduation secondaire
-			SVG_graduation(mon_svg,100,longueur_pas2,750,taille=5,y=50,color='black',width=2)
+			SVG_graduation(mon_svg,100,longueur_pas2,750,5,50,'black',2)
 			// Graduation principale
-			SVG_graduation(mon_svg,100,longueur_pas1,750,taille=10,y=50,color='black',width=5)
+			SVG_graduation(mon_svg,100,longueur_pas1,750,10,50,'black',5)
 			SVG_tracer_flecheH(mon_svg,750,50)
 			// Nombres visibles
 			SVG_label(mon_svg,[[string_nombre(origine),100,50]],2,'black',1);
